% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytreaties-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements datacube documentation}
\format{
The agreements datacube is a list that contains the
following 9 datasets: GHHR, GPTAD, HEIDI, IEADB, LABPTA, TOTA, TREND, WHO,
HUGGO.
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and
\code{manydata::compare_dimensions()} functions.
\describe{
\item{GHHR: }{A dataset with 149 observations and the following
8 variables: manyID, treatyID, Title, Begin, Region, LegalStatus, Lineage,
ghhrID.}
\item{GPTAD: }{A dataset with 340 observations and the following
10 variables: manyID, treatyID, Title, Begin, AgreementType, DocType,
GeogArea, Signature, Force, gptadID.}
\item{HEIDI: }{A dataset with 2280 observations and the following
7 variables: manyID, treatyID, Title, Begin, Signature, Lineage, heidiID.}
\item{IEADB: }{A dataset with 3667 observations and the following
10 variables: manyID, treatyID, Title, Begin, DocType, AgreementType,
Signature, Force, Lineage, ieadbID.}
\item{LABPTA: }{A dataset with 483 observations and the following
7 variables: manyID, treatyID, Title, Begin, Signature, Force, labptaID.}
\item{TOTA: }{A dataset with 442 observations and the following
7 variables: manyID, treatyID, Title, Begin, Signature, Force, totaID.}
\item{TREND: }{A dataset with 710 observations and the following
7 variables: manyID, treatyID, Title, Begin, Signature, Force, trendID.}
\item{WHO: }{A dataset with 114 observations and the following
6 variables: manyID, treatyID, Title, Begin, Organisation, Topic.}
\item{HUGGO: }{A dataset with 4837 observations and the following
59 variables: manyID, treatyID, Title, Begin, End, Signature, Force,
TreatyText, url, Domain, AgreementType, DocType, GeogArea, gengID, ieaID,
ecolexID, Parties, verified, DocValidUntilDate, Notes, Download, MEA_type,
Ambit, Region, subject_ecolex, subject_iea, Keywords, Lineage, Sequence,
AdoptedIn, Languages, Appendices, Depository, DepositoryURL, Published,
Website1, Website2, Secretariat, SecretariatURL, UNEP, Supersedes,
References, EnabledBy, AmendedBy, Lit, Data, Coded, Abstract, Language,
Orig_noneng_title, match, Citation, Formal, Dataset, Source,
Health_as_primary_intent, Comments, Topic, Coder.}
}
}
\source{
\itemize{
\item{GHHR: }{
B. M. Meier, O. A. Cabrera, A. Ayala, et al.
“Bridging international law and rights-based litigation:mapping health-related rights through the development of the global health and human rights database”.
\emph{Health & Human Rights.} 14 (2012), p. 20.}
\item{GPTAD: }{
World Bank Group. \emph{Global Preferential Trade Agreement Database (GPTAD)}.
Online database. publisher: World Bank Group. 2014.
\url{https://wits.worldbank.org/gptad/library.aspx}.}
\item{HEIDI: }{
J-F. Morin, and C. Blouin. "How environmental treaties contribute to global health governance".
\emph{Globalization and health} 15.1 (2019), pp. 1-8.
}
\item{IEADB: }{
R. B. Mitchell et al. "What we know (and could know) about international environmental agreements".
\emph{Global Environmental Politics} 20.1 (2020), pp. 103-121.
}
\item{LABPTA: }{
D. Raess, A. Dür, and D. Sari.
“Protecting labor rights in preferential trade agreements: The role of trade unions, left governments, and skilled labor”.
In: \emph{The Review of International Organizations} 2.13(2018), pp. 143-162.
DOI: 10.1007/s11558-018-9301-z.}
\item{TOTA: }{
W. Alschner, J. Seiermann, and D. Skougarevskiy.
\emph{Text-as-data analysis of preferential tradeagreements: Mapping the PTA landscape}.
2017. \url{https://github.com/mappingtreaties/tota.git}.}
\item{TREND: }{
J. Morin, A. Dür, and L. Lechner.
“Mapping the trade and environment nexus: Insights from a newdataset”.
In: \emph{Global Environmental Politics} 18.1 (2018), pp. 122-139.}
\item{WHO: }{
World Health Organization.
\emph{WHO MiNDbank: More Inclusiveness Needed in Disability and Development}.
Online database. publisher: World Health Organization. 2024.
\url{https://extranet.who.int/mindbank/}.}
\item{HUGGO: }{
J. Hollway. International agreements for manydata. 2021.
}
}
}
\usage{
agreements
}
\description{
agreements datacube documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $GHHR
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             149|     0     |       0       |
#> |   treatyID    |  character  |             149|     0     |       0       |
#> |     Title     |  character  |             149|     0     |       0       |
#> |     Begin     |    mdate    |             149|     3     |     2.01      |
#> |    Region     |  character  |             149|     0     |       0       |
#> |  LegalStatus  |  character  |             149|     0     |       0       |
#> |    Lineage    |  character  |             149|    52     |     34.9      |
#> |    ghhrID     |  character  |             149|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $GPTAD
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             340|     0     |       0       |
#> |   treatyID    |  character  |             340|     0     |       0       |
#> |     Title     |  character  |             340|     0     |       0       |
#> |     Begin     |    mdate    |             340|    30     |     8.82      |
#> | AgreementType |  character  |             340|     0     |       0       |
#> |    DocType    |  character  |             340|     0     |       0       |
#> |   GeogArea    |  character  |             340|    75     |     22.06     |
#> |   Signature   |    mdate    |             340|    53     |     15.59     |
#> |     Force     |    mdate    |             340|    63     |     18.53     |
#> |    gptadID    |  character  |             340|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HEIDI
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2280|     0     |       0       |
#> |   treatyID    |  character  |            2280|     0     |       0       |
#> |     Title     |  character  |            2280|     0     |       0       |
#> |     Begin     |    mdate    |            2280|    27     |     1.18      |
#> |   Signature   |    mdate    |            2280|    27     |     1.18      |
#> |    Lineage    |  character  |            2280|     0     |       0       |
#> |    heidiID    |   numeric   |            2280|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $IEADB
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            3667|     0     |       0       |
#> |   treatyID    |  character  |            3667|     0     |       0       |
#> |     Title     |  character  |            3667|     0     |       0       |
#> |     Begin     |    mdate    |            3667|     0     |       0       |
#> |    DocType    |  character  |            3667|     0     |       0       |
#> | AgreementType |  character  |            3667|     0     |       0       |
#> |   Signature   |    mdate    |            3667|    65     |     1.77      |
#> |     Force     |    mdate    |            3667|    868    |     23.67     |
#> |    Lineage    |  character  |            3667|     0     |       0       |
#> |    ieadbID    |  character  |            3667|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $LABPTA
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             483|     0     |       0       |
#> |   treatyID    |  character  |             483|     0     |       0       |
#> |     Title     |  character  |             483|     0     |       0       |
#> |     Begin     |    mdate    |             483|     0     |       0       |
#> |   Signature   |    mdate    |             483|     0     |       0       |
#> |     Force     |    mdate    |             483|     0     |       0       |
#> |   labptaID    |  character  |             483|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $TOTA
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             442|     0     |       0       |
#> |   treatyID    |  character  |             442|     0     |       0       |
#> |     Title     |  character  |             442|     0     |       0       |
#> |     Begin     |    mdate    |             442|     0     |       0       |
#> |   Signature   |    mdate    |             442|     0     |       0       |
#> |     Force     |    mdate    |             442|     3     |     0.68      |
#> |    totaID     |  character  |             442|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $TREND
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             710|     0     |       0       |
#> |   treatyID    |  character  |             710|     0     |       0       |
#> |     Title     |  character  |             710|     0     |       0       |
#> |     Begin     |    mdate    |             710|     5     |      0.7      |
#> |   Signature   |    mdate    |             710|     5     |      0.7      |
#> |     Force     |    mdate    |             710|     5     |      0.7      |
#> |    trendID    |  character  |             710|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $WHO
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             114|     0     |       0       |
#> |   treatyID    |  character  |             114|     0     |       0       |
#> |     Title     |  character  |             114|     0     |       0       |
#> |     Begin     |    mdate    |             114|     0     |       0       |
#> | Organisation  |  character  |             114|     0     |       0       |
#> |     Topic     |  character  |             114|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO
#> -------------------------------------------------------------------------------------
#> |      Column Name       |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> -------------------------------------------------------------------------------------
#> |         manyID         |  character  |            4837|     2     |     0.04      |
#> |        treatyID        |  character  |            4837|     2     |     0.04      |
#> |         Title          |  character  |            4837|     0     |       0       |
#> |         Begin          |    mdate    |            4837|    57     |     1.18      |
#> |          End           |    mdate    |            4837|   3156    |     65.25     |
#> |       Signature        |    mdate    |            4837|    57     |     1.18      |
#> |         Force          |    mdate    |            4837|   1816    |     37.54     |
#> |       TreatyText       |  character  |            4837|     0     |       0       |
#> |          url           |  character  |            4837|   2796    |     57.8      |
#> |         Domain         |  character  |            4837|     0     |       0       |
#> |     AgreementType      |  character  |            4837|   2132    |     44.08     |
#> |        DocType         |  character  |            4837|   2096    |     43.33     |
#> |        GeogArea        |  character  |            4837|   2697    |     55.76     |
#> |         gengID         |  character  |            4837|   1054    |     21.79     |
#> |         ieaID          |  character  |            4837|   4142    |     85.63     |
#> |        ecolexID        |  character  |            4837|   4360    |     90.14     |
#> |        Parties         |  character  |            4837|   2567    |     53.07     |
#> |        verified        |  character  |            4837|   4789    |     99.01     |
#> |   DocValidUntilDate    |  character  |            4837|   4803    |     99.3      |
#> |         Notes          |  character  |            4837|   4790    |     99.03     |
#> |        Download        |  character  |            4837|   3939    |     81.43     |
#> |        MEA_type        |  character  |            4837|   4142    |     85.63     |
#> |         Ambit          |  character  |            4837|   4419    |     91.36     |
#> |         Region         |  character  |            4837|   4591    |     94.91     |
#> |     subject_ecolex     |  character  |            4837|   4361    |     90.16     |
#> |      subject_iea       |  character  |            4837|   4142    |     85.63     |
#> |        Keywords        |  character  |            4837|   4431    |     91.61     |
#> |        Lineage         |  character  |            4837|   4142    |     85.63     |
#> |        Sequence        |  character  |            4837|   4143    |     85.65     |
#> |       AdoptedIn        |  character  |            4837|   4411    |     91.19     |
#> |       Languages        |  character  |            4837|   4361    |     90.16     |
#> |       Appendices       |  character  |            4837|   4802    |     99.28     |
#> |       Depository       |  character  |            4837|   4451    |     92.02     |
#> |     DepositoryURL      |  character  |            4837|   4821    |     99.67     |
#> |       Published        |  character  |            4837|   4576    |     94.6      |
#> |        Website1        |  character  |            4837|   4456    |     92.12     |
#> |        Website2        |  character  |            4837|   4813    |     99.5      |
#> |      Secretariat       |  character  |            4837|   4143    |     85.65     |
#> |     SecretariatURL     |  character  |            4837|   4143    |     85.65     |
#> |          UNEP          |  character  |            4837|   4636    |     95.84     |
#> |       Supersedes       |  character  |            4837|   4820    |     99.65     |
#> |       References       |  character  |            4837|   4802    |     99.28     |
#> |       EnabledBy        |  character  |            4837|   4730    |     97.79     |
#> |       AmendedBy        |  character  |            4837|   4746    |     98.12     |
#> |          Lit           |  character  |            4837|   4292    |     88.73     |
#> |          Data          |  character  |            4837|   4300    |     88.9      |
#> |         Coded          |  character  |            4837|   4539    |     93.84     |
#> |        Abstract        |  character  |            4837|   4811    |     99.46     |
#> |        Language        |  character  |            4837|   3488    |     72.11     |
#> |   Orig_noneng_title    |  character  |            4837|   4781    |     98.84     |
#> |         match          |  character  |            4837|   4781    |     98.84     |
#> |        Citation        |  character  |            4837|   4836    |     99.98     |
#> |         Formal         |  character  |            4837|   4608    |     95.27     |
#> |        Dataset         |  character  |            4837|   4608    |     95.27     |
#> |         Source         |  character  |            4837|   4609    |     95.29     |
#> |Health_as_primary_intent|  character  |            4837|   4608    |     95.27     |
#> |        Comments        |  character  |            4837|   4814    |     99.52     |
#> |         Topic          |  character  |            4837|   4608    |     95.27     |
#> |         Coder          |  character  |            4837|     0     |       0       |
#> -------------------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{URL}{

\itemize{
\item{GHHR: }{
\url{https://www.globalhealthrights.org/instruments/instrument-region/}
}
\item{GPTAD: }{
\url{https://wits.worldbank.org/gptad/library.aspx}
}
\item{HEIDI: }{
\url{https://www.chaire-epi.ulaval.ca/en/data/heidi}
}
\item{IEADB: }{
\url{https://iea.uoregon.edu/base-agreement-list}
}
\item{LABPTA: }{
\url{https://doi.org/10.1007/s11558-018-9301-z}
}
\item{TOTA: }{
\url{https://github.com/mappingtreaties/tota.git}
}
\item{TREND: }{
\url{http://www.chaire-epi.ulaval.ca/en/trend}
}
\item{WHO: }{
\url{https://www.mindbank.info/collection/un_who_resolutions/all?page=all}
}
\item{HUGGO: }{Hand-coded data by the GGO team.}
}
}

\section{Mapping}{

\itemize{
\item{GHHR: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   Year of adoption \tab Begin \cr
}


}
\item{GPTAD: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   Common.Name \tab Title \cr
   Date.of.Signature \tab Signature \cr
   Date.of.Entry.into.Force \tab Force \cr
   Type \tab DocType \cr
   Type \tab AgreementType \cr
   Type \tab GeogArea \cr
}


}
\item{HEIDI: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   ID \tab heidiID \cr
   'Name.of.the.agreement' \tab Title \cr
   signature.date \tab Signature \cr
}


}
\item{IEADB: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   'IEA# (click for add'l info)' \tab ieadbID \cr
   'Treaty Name' \tab Title \cr
   'Signature Date' \tab Signature \cr
   'Date IEA entered into force' \tab Force \cr
   'Agreement Type (level 2)' \tab AgreementType \cr
   Inclusion \tab DocType \cr
}


}
\item{LABPTA: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   Number \tab labptaID \cr
   Name \tab Title \cr
   year \tab Signature \cr
   year \tab Force \cr
}


}
\item{TOTA: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   name \tab Title \cr
   date_signed \tab Signature \cr
   date_into_force \tab Force \cr
}


}
\item{TREND: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   Trade.Agreement \tab Title \cr
   Year \tab Signature \cr
   Year \tab Force \cr
}


}
\item{WHO: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   NA \tab NA \cr
}


}
\item{HUGGO: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   NA \tab NA \cr
}


}
}
}

\keyword{datasets}
