% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_lineage.R
\name{code_area}
\alias{code_area}
\title{Code areas from agreement titles}
\usage{
code_area(title, return = c("full", "short"))
}
\arguments{
\item{title}{Treaty titles}

\item{type}{Issue-type of agreement, either environment or health}
}
\value{
The area taken from agreement title
}
\description{
Code areas from agreement titles
}
\examples{
\dontrun{
code_area(HUGGO$Title)
}
}
