% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_lineage.R
\name{code_domain}
\alias{code_domain}
\title{Code domains from agreement titles}
\usage{
code_domain(title, type = c("environment", "health"))
}
\arguments{
\item{title}{Treaty titles}

\item{type}{Issue-type of agreement, either environment or health}
}
\value{
The domain taken from agreement title
}
\description{
Code domains from agreement titles
}
\examples{
\dontrun{
title <- sample(manyenviron::agreements$IEADB$Title, 30)
code_domain(title)
}
}
