% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_lineage.R
\name{code_location}
\alias{code_location}
\title{Code locations from agreement titles}
\usage{
code_location(title)
}
\arguments{
\item{title}{Treaty titles}
}
\value{
The location taken from agreement title
}
\description{
Code locations from agreement titles
}
\examples{
\dontrun{
code_location(HUGGO$Title)
}
}
