% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_agreements.R
\name{generate_}
\alias{generate_}
\alias{generate_agreements}
\title{Generate agreements}
\usage{
generate_agreements(n = 10, treaty_type = "any")
}
\arguments{
\item{n}{Integer number of fictional agreements to generate.}

\item{treaty_type}{What type of agreements would you like to generate?
By default, "any" agreements.
Other options include "bilaterals" or "multilaterals".}
}
\value{
String vector of fictional agreement names and/or memberships
to agreements.
}
\description{
Generates a list of fictional agreements or a
list memberships to agreements.
}
\examples{
\donttest{
generate_agreements(12)
generate_agreements(12, treaty_type = "bilaterals")
generate_agreements(12, treaty_type = "multilaterals")
}
}
