% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytreaties-package.R
\docType{package}
\name{manytreaties-package}
\alias{manytreaties}
\alias{manytreaties-package}
\title{manytreaties: Many international agreements and their parties}
\description{
Ensemble of data on international environmental, trade, and health agreements, and states membership in them for research use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/globalgov/manytreaties}
  \item Report bugs at \url{https://github.com/globalgov/manytreaties/issues}
}

}
\author{
\strong{Maintainer}: James Hollway \email{james.hollway@graduateinstitute.ch} (\href{https://orcid.org/0000-0002-8361-9647}{ORCID}) (IHEID) [contributor]

Other contributors:
\itemize{
  \item Jael Tan (\href{https://orcid.org/0000-0002-6234-9764}{ORCID}) (IHEID) [contributor]
}

}
\keyword{internal}
