% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytreaties-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships datacube documentation}
\format{
The memberships datacube is a list that contains the
following 4 datasets: GPTAD_MEM, IEADB_MEM, TFDD_MEM, HUGGO_MEM.
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
\describe{
\item{GPTAD_MEM: }{A dataset with 2192 observations and the following
9 variables: manyID, treatyID, Title, Begin, stateID, Signature, Force,
StateName, gptadID.}
\item{IEADB_MEM: }{A dataset with 15463 observations and the following
12 variables: manyID, treatyID, Title, Begin, stateID, End, stateSignature,
Signature, Rat, Force, DocType, ieadbID.}
\item{TFDD_MEM: }{A dataset with 1832 observations and the following
8 variables: manyID, treatyID, Title, Begin, stateID, Signature, tfddID, Memberships.}
\item{HUGGO_MEM: }{A dataset with 91664 observations and the following
41 variables: manyID, treatyID, Title, Begin, stateID, StateName,
StateSignature, StateRatification, StateForce, StateEnd, Rat=Notif,
Accession, Succession, Signature, Force, End, gengID, url, ieaID, ecolexID,
StateForce2, StateForce3, StateEnd2, Term, Comments, Deposit, obsolete,
ProvisionalApp, Reservation, Notes, stateForce_ecolex, stateForce_iea,
Consent, Acceptance, Orig_noneng_title, match, gptadID, destaID, Title2,
Domain, Coder.}
}
}
\source{
\itemize{
\item{GPTAD_MEM: }{
W. B. Group.
\emph{Global Preferential Trade Agreement Database (GPTAD)}.
\url{https://wits.worldbank.org/gptad/library.aspx}. 2013.}
\item{IEADB_MEM: }{
R. B. Mitchell, L. B. Andonova, M. Axelrod, et al. “What we know (and could know) about internationalenvironmental agreements”.
\emph{Global Environmental Politics} 20.1 (2020), pp. 103-121.}
\item{TFDD_MEM: }{
O. College of Earth and O. S. U. Atmospheric Science.
\emph{Product of the Transboundary Freshwater DisputeDatabase}. \url{http://transboundarywaters.science.oregonstate.edu.}. 2021.}
\link{1} D. Bhugra, S. Pathare, R. Joshi, et al. “A review of mental health policies from Commonwealthcountries”. In: \emph{International Journal of Social Psychiatry} 64.1 (2018), pp. 3-8.\link{2} B. M. Meier, O. A. Cabrera, A. Ayala, et al. “Bridging international law and rights-basedlitigation: mapping health-related rights through the development of the global health and humanrights database”. In: \emph{Health & Hum. Rts.} 14 (2012), p. 20.}
}
}
\usage{
memberships
}
\description{
memberships datacube documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $GPTAD_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2192|     0     |       0       |
#> |   treatyID    |  character  |            2192|     0     |       0       |
#> |     Title     |  character  |            2192|     0     |       0       |
#> |     Begin     |    mdate    |            2192|    99     |     4.52      |
#> |    stateID    |  character  |            2192|    55     |     2.51      |
#> |   Signature   |    mdate    |            2192|    310    |     14.14     |
#> |     Force     |    mdate    |            2192|    285    |      13       |
#> |   StateName   |  character  |            2192|    36     |     1.64      |
#> |    gptadID    |  character  |            2192|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $IEADB_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |           15463|     0     |       0       |
#> |   treatyID    |  character  |           15463|     0     |       0       |
#> |     Title     |  character  |           15463|     0     |       0       |
#> |     Begin     |    mdate    |           15463|     0     |       0       |
#> |    stateID    |  character  |           15463|     0     |       0       |
#> |      End      |    mdate    |           15463|   14593   |     94.37     |
#> |stateSignature |    mdate    |           15463|   5797    |     37.49     |
#> |   Signature   |    mdate    |           15463|     0     |       0       |
#> |      Rat      |    mdate    |           15463|   5485    |     35.47     |
#> |     Force     |    mdate    |           15463|     0     |       0       |
#> |    DocType    |  character  |           15463|     0     |       0       |
#> |    ieadbID    |   numeric   |           15463|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $TFDD_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            1832|     0     |       0       |
#> |   treatyID    |  character  |            1832|     0     |       0       |
#> |     Title     |  character  |            1832|     1     |     0.05      |
#> |     Begin     |    mdate    |            1832|    102    |     5.57      |
#> |    stateID    |  character  |            1832|     1     |     0.05      |
#> |   Signature   |    mdate    |            1832|    102    |     5.57      |
#> |    tfddID     |   numeric   |            1832|     1     |     0.05      |
#> |  Memberships  |  character  |            1832|    805    |     43.94     |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO_MEM
#> ------------------------------------------------------------------------------
#> |   Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ------------------------------------------------------------------------------
#> |     manyID      |  character  |           91664|     0     |       0       |
#> |    treatyID     |  character  |           91664|     0     |       0       |
#> |      Title      |  character  |           91664|     0     |       0       |
#> |      Begin      |    mdate    |           91664|    12     |     0.01      |
#> |     stateID     |  character  |           91664|     0     |       0       |
#> |    StateName    |  character  |           91664|   50485   |     55.08     |
#> | StateSignature  |    mdate    |           91664|   17944   |     19.58     |
#> |StateRatification|    mdate    |           91664|   63905   |     69.72     |
#> |   StateForce    |    mdate    |           91664|   50866   |     55.49     |
#> |    StateEnd     |    mdate    |           91664|   33488   |     36.53     |
#> |    Rat=Notif    |  character  |           91664|   90447   |     98.67     |
#> |    Accession    |    mdate    |           91664|   77166   |     84.18     |
#> |   Succession    |  character  |           91664|     0     |       0       |
#> |    Signature    |    mdate    |           91664|    16     |     0.02      |
#> |      Force      |    mdate    |           91664|   34564   |     37.71     |
#> |       End       |    mdate    |           91664|   38809   |     42.34     |
#> |     gengID      |  character  |           91664|   40705   |     44.41     |
#> |       url       |  character  |           91664|   90814   |     99.07     |
#> |      ieaID      |  character  |           91664|   49313   |     53.8      |
#> |    ecolexID     |  character  |           91664|   71665   |     78.18     |
#> |   StateForce2   |    mdate    |           91664|   84977   |     92.7      |
#> |   StateForce3   |    mdate    |           91664|   87974   |     95.97     |
#> |    StateEnd2    |    mdate    |           91664|   87968   |     95.97     |
#> |      Term       |    mdate    |           91664|   90486   |     98.71     |
#> |    Comments     |  character  |           91664|   88384   |     96.42     |
#> |     Deposit     |  character  |           91664|   90014   |     98.2      |
#> |    obsolete     |  character  |           91664|   79089   |     86.28     |
#> | ProvisionalApp  |  character  |           91664|   90892   |     99.16     |
#> |   Reservation   |  character  |           91664|   90908   |     99.18     |
#> |      Notes      |  character  |           91664|   91478   |     99.8      |
#> |stateForce_ecolex|    mdate    |           91664|   73147   |     79.8      |
#> | stateForce_iea  |    mdate    |           91664|   73147   |     79.8      |
#> |     Consent     |  character  |           91664|   91440   |     99.76     |
#> |   Acceptance    |  character  |           91664|   91139   |     99.43     |
#> |Orig_noneng_title|  character  |           91664|   91660   |      100      |
#> |      match      |  character  |           91664|   91660   |      100      |
#> |     gptadID     |  character  |           91664|   89495   |     97.63     |
#> |     destaID     |  character  |           91664|   86612   |     94.49     |
#> |     Title2      |  character  |           91664|   91187   |     99.48     |
#> |     Domain      |  character  |           91664|     0     |       0       |
#> |      Coder      |  character  |           91664|    15     |     0.02      |
#> ------------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{URL}{

\itemize{
\item{GPTAD_MEM: }{
\url{https://wits.worldbank.org/gptad/library.aspx}
}
\item{IEADB_MEM: }{
\url{https://iea.uoregon.edu/country-members}
}
\item{TFDD_MEM: }{
\url{https://transboundarywaters.science.oregonstate.edu/}
}
\item{HUGGO_MEM: }{Hand-coded data by the GGO team.}
}
}

\section{Mapping}{

\itemize{
\item{GPTAD_MEM: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   Membership \tab StateName \cr
   Common.Name \tab Title \cr
   Date.of.Signature \tab Signature \cr
   Date.of.Entry.into.Force \tab Force \cr
}


}
\item{IEADB_MEM: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   country \tab stateID \cr
   treatyname \tab Title \cr
   tsig \tab Signature \cr
   csig \tab stateSignature \cr
   crat \tab Rat \cr
   tterm \tab End \cr
   ceif3 \tab Force \cr
   ceif4 \tab Force2 \cr
   mitch_id \tab ieadbID \cr
   inclusion \tab DocType \cr
}


}
\item{TFDD_MEM: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   DateSigned \tab Signature \cr
   '2016Update ID' \tab tfddID \cr
   CCODE \tab stateID \cr
   DocumentName \tab Title \cr
   Signatories \tab Memberships \cr
}


}
\item{HUGGO_MEM: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   NA \tab NA \cr
}


}
}
}

\keyword{datasets}
