% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytreaties-parties.R
\docType{data}
\name{parties}
\alias{parties}
\title{parties datacube}
\format{
\describe{
\item{HUGGO \if{html}{\out{<img src="https://mirrors.creativecommons.org/presskit/buttons/88x31/png/by.png" height="12"/>}}:}{A dataset with 91,729
observations and 31 variables:
stateID, treatyID, Begin, End, manyID, Title, StateSignature, StateRatification, StateForce, StateEnd, Rat=Notif, Accession, Succession, Signature, Force, url, StateForce2, StateForce3, ..., match, and Coder.}
\item{IEADB: }{A dataset with 15,463
observations and 11 variables:
stateID, treatyID, Begin, End, manyID, Title, stateSignature, Signature, Rat, Force, and ieadbID.}
\item{GPTAD \if{html}{\out{<img src="https://mirrors.creativecommons.org/presskit/buttons/88x31/png/by.png" height="12"/>}}:}{A dataset with 2,192
observations and 7 variables:
stateID, treatyID, Begin, manyID, Title, Signature, and Force.}
\item{TFDD \if{html}{\out{<img src="https://mirrors.creativecommons.org/presskit/buttons/88x31/png/cc-zero.png" height="12"/>}}:}{A dataset with 1,832
observations and 7 variables:
stateID, treatyID, Begin, manyID, Title, Signature, and tfddID.}
}
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
}
\source{
\itemize{
\item Hollway, James, Henrique Sposito, and Jael Tan. 2021. International agreements for manydata.
\item Mitchell, Ron B. et al. 2020. 'What we know (and could know) about international environmental agreements'. \emph{Global Environmental Politics} 20.1, pp. 103-121.
\item World Bank Group. 2014. \emph{Global Preferential Trade Agreement Database (GPTAD)}. Online database. World Bank Group.
\item Oregon College of Earth and Oregon State University Atmospheric Science. 2021. \emph{Product of the Transboundary Freshwater Dispute Database}.
}
}
\usage{
parties
}
\description{
The \code{manytreaties::parties} datacube is a list containing 4 datasets: HUGGO, IEADB, GPTAD, and TFDD.
It is a work-in-progress, so please let us know if you have any comments or suggestions.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $HUGGO
#> ------------------------------------------------------------
#> |    Variable     |  Class  |  Obs  |  Missing  |  Miss \%  |
#> ------------------------------------------------------------
#> |stateID          |character|  91729|          0|         0|
#> |treatyID         |character|  91729|          0|         0|
#> |Begin            |mdate    |  91717|         12|      0.01|
#> |End              |mdate    |  52937|      38792|     42.29|
#> |manyID           |character|  91729|          0|         0|
#> |Title            |character|  91729|          0|         0|
#> |StateSignature   |mdate    |  73785|      17944|     19.56|
#> |StateRatification|mdate    |  27769|      63960|     69.73|
#> |StateForce       |mdate    |  40878|      50851|     55.44|
#> |StateEnd         |mdate    |  58258|      33471|     36.49|
#> |Rat=Notif        |character|   1227|      90502|     98.66|
#> |Accession        |mdate    |  14499|      77230|     84.19|
#> |Succession       |character|  91729|          0|         0|
#> |Signature        |mdate    |  91713|         16|      0.02|
#> |Force            |mdate    |  57180|      34549|     37.66|
#> |url              |character|    931|      90798|     98.99|
#> |StateForce2      |mdate    |   6687|      85042|     92.71|
#> |StateForce3      |mdate    |   3690|      88039|     95.98|
#> |StateEnd2        |mdate    |   3696|      88033|     95.97|
#> |Term             |mdate    |   1178|      90551|     98.72|
#> |Comments         |character|   3280|      88449|     96.42|
#> |obsolete         |character|  12575|      79154|     86.29|
#> |ProvisionalApp   |character|    772|      90957|     99.16|
#> |Reservation      |character|    756|      90973|     99.18|
#> |Notes            |character|    186|      91543|      99.8|
#> |stateForce_ecolex|mdate    |  18517|      73212|     79.81|
#> |stateForce_iea   |mdate    |  18517|      73212|     79.81|
#> |Consent          |character|    224|      91505|     99.76|
#> |Acceptance       |character|    525|      91204|     99.43|
#> |match            |character|      4|      91725|       100|
#> |Coder            |character|  91714|         15|      0.02|
#> ------------------------------------------------------------
#> 
#> 
#> $IEADB
#> ---------------------------------------------------------
#> |   Variable   |  Class  |  Obs  |  Missing  |  Miss \%  |
#> ---------------------------------------------------------
#> |stateID       |character|  15463|          0|         0|
#> |treatyID      |character|  15463|          0|         0|
#> |Begin         |mdate    |  15463|          0|         0|
#> |End           |mdate    |    870|      14593|     94.37|
#> |manyID        |character|  15463|          0|         0|
#> |Title         |character|  15463|          0|         0|
#> |stateSignature|mdate    |   9666|       5797|     37.49|
#> |Signature     |mdate    |  15463|          0|         0|
#> |Rat           |mdate    |   9978|       5485|     35.47|
#> |Force         |mdate    |  15463|          0|         0|
#> |ieadbID       |numeric  |  15463|          0|         0|
#> ---------------------------------------------------------
#> 
#> 
#> $GPTAD
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |stateID     |character|   2137|         55|      2.51|
#> |treatyID    |character|   2192|          0|         0|
#> |Begin       |mdate    |   2093|         99|      4.52|
#> |manyID      |character|   2192|          0|         0|
#> |Title       |character|   2192|          0|         0|
#> |Signature   |mdate    |   1882|        310|     14.14|
#> |Force       |mdate    |   1907|        285|        13|
#> -------------------------------------------------------
#> 
#> 
#> $TFDD
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |stateID     |character|   1831|          1|      0.05|
#> |treatyID    |character|   1832|          0|         0|
#> |Begin       |mdate    |   1730|        102|      5.57|
#> |manyID      |character|   1832|          0|         0|
#> |Title       |character|   1831|          1|      0.05|
#> |Signature   |mdate    |   1730|        102|      5.57|
#> |tfddID      |numeric  |   1831|          1|      0.05|
#> -------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{Mapping}{
\tabular{llll}{
   \emph{manytreaties} \tab \emph{IEADB} \tab \emph{GPTAD} \tab \emph{TFDD} \cr
   stateID \tab country \tab  \tab CCODE \cr
   Signature \tab csig \tab Date.of.Signature \tab DateSigned \cr
   Rat \tab crat \tab  \tab  \cr
   Force \tab ceif3 \tab Date.of.Entry.into.Force \tab  \cr
   Force2 \tab ceif4 \tab  \tab  \cr
   End \tab tterm \tab  \tab  \cr
   Title \tab treatyname \tab Common.Name \tab DocumentName \cr
   Memberships \tab  \tab StateName \tab Signatories \cr
   tfddID \tab  \tab  \tab 2016Update ID \cr
   mitchID \tab ieadbID \tab  \tab  \cr
   AgreementType \tab  \tab  \tab  \cr
   DocType \tab inclusion \tab  \tab  \cr
}
}

\keyword{datasets}
