% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytreaties-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements datacube documentation}
\format{
\describe{
\item{HUGGO \if{html}{\out{<img src="https://mirrors.creativecommons.org/presskit/buttons/88x31/png/by.png" height="12"/>}}:}{A dataset with 4,829
observations and 28 variables:
manyID, treatyID, Title, Begin, End, Signature, Force, Term, Domain, AgreementType, Ambit, Region, Keywords, Lineage, AdoptedIn, Depository, Published, Secretariat, ..., gengID, and Coder.}
\item{IEADB: }{A dataset with 3,667
observations and 10 variables:
manyID, treatyID, Title, Begin, Ambit, AgreementType, Signature, Force, Lineage, and mitchID.}
\item{HEIDI: }{A dataset with 2,280
observations and 7 variables:
manyID, treatyID, Title, Begin, Signature, Lineage, and heidiID.}
\item{TFDD: }{A dataset with 784
observations and 14 variables:
manyID, Title, Begin, Signature, AgreementType, Ambit, GeogArea, Water_as_consumable_rsc, NotTreaty, Issue, Basin, Lineage, tfddID, and treatyID.}
\item{TREND: }{A dataset with 710
observations and 7 variables:
manyID, treatyID, Title, Begin, Signature, Force, and trendID.}
\item{LABPTA \if{html}{\out{<img src="https://mirrors.creativecommons.org/presskit/buttons/88x31/png/by.png" height="12"/>}}:}{A dataset with 483
observations and 7 variables:
manyID, treatyID, Title, Begin, Signature, Force, and labptaID.}
\item{TOTA \if{html}{\out{<img src="https://mirrors.creativecommons.org/presskit/buttons/88x31/png/by.png" height="12"/>}}:}{A dataset with 442
observations and 7 variables:
manyID, treatyID, Title, Begin, Signature, Force, and totaID.}
\item{GPTAD \if{html}{\out{<img src="https://mirrors.creativecommons.org/presskit/buttons/88x31/png/by.png" height="12"/>}}:}{A dataset with 340
observations and 10 variables:
manyID, treatyID, Title, Begin, AgreementType, Ambit, GeogArea, Signature, Force, and gptadID.}
}
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
}
\source{
\itemize{
\item Hollway, James, Henrique Sposito, and Jael Tan. 2021. International agreements for manydata.
\item Mitchell, Ron B. et al. 2020. 'What we know (and could know) about international environmental agreements'. \emph{Global Environmental Politics} 20.1, pp. 103-121.
\item Morin, Jean-Frederic, and Chantal Blouin. 2019. 'How environmental treaties contribute to global health governance'. \emph{Globalization and health} 15.1, pp. 1-8.
\item Oregon College of Earth and Oregon State University Atmospheric Science. 2021. \emph{Product of the Transboundary Freshwater Dispute Database}.
\item Morin, Jean-Frederic, Andreas Duer, and Lisa Lechner. 2018. 'Mapping the trade and environment nexus: Insights from a new dataset'. \emph{Global Environmental Politics} 18.1, pp. 122-139.
\item Raess, Damien, Andreas Duer, and D. Sari. 2018. 'Protecting labor rights in preferential trade agreements: The role of trade unions, left governments, and skilled labor'. \emph{The Review of International Organizations} 2.13, pp. 143-162.
\item Alschner, Wolfgang, Julia Seiermann, and Dmitriy Skougarevskiy. 2017. 'Text-as-data analysis of preferential trade agreements: Mapping the PTA landscape'. UNCTAD Research Paper No. 5.
\item World Bank Group. 2014. \emph{Global Preferential Trade Agreement Database (GPTAD)}. Online database. World Bank Group.
}
}
\usage{
agreements
}
\description{
The \code{manytreaties::agreements} datacube is a list containing 8 datasets: HUGGO, IEADB, HEIDI, TFDD, TREND, LABPTA, TOTA, and GPTAD.
It is a work-in-progress, so please let us know if you have any comments or suggestions.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $HUGGO
#> --------------------------------------------------------
#> |  Variable   |  Class  |  Obs  |  Missing  |  Miss \%  |
#> --------------------------------------------------------
#> |manyID       |character|   4829|          0|         0|
#> |treatyID     |character|   4829|          0|         0|
#> |Title        |character|   4829|          0|         0|
#> |Begin        |mdate    |   4772|         57|      1.18|
#> |End          |mdate    |   1681|       3148|     65.19|
#> |Signature    |mdate    |   4772|         57|      1.18|
#> |Force        |mdate    |   3016|       1813|     37.54|
#> |Term         |character|     32|       4797|     99.34|
#> |Domain       |character|   4803|         26|      0.54|
#> |AgreementType|character|   4004|        825|     17.08|
#> |Ambit        |character|   3551|       1278|     26.47|
#> |Region       |character|    246|       4583|     94.91|
#> |Keywords     |character|    406|       4423|     91.59|
#> |Lineage      |character|    695|       4134|     85.61|
#> |AdoptedIn    |character|    426|       4403|     91.18|
#> |Depository   |character|    386|       4443|     92.01|
#> |Published    |character|    261|       4568|      94.6|
#> |Secretariat  |character|    694|       4135|     85.63|
#> |Website      |character|    843|       3986|     82.54|
#> |OrigTitle    |character|     56|       4773|     98.84|
#> |Language     |character|   1349|       3480|     72.06|
#> |Languages    |character|    476|       4353|     90.14|
#> |match        |character|     56|       4773|     98.84|
#> |Text         |character|   4829|          0|         0|
#> |TextURL      |character|   2047|       2782|     57.61|
#> |Comments     |character|    358|       4471|     92.59|
#> |gengID       |character|   3775|       1054|     21.83|
#> |Coder        |character|   4829|          0|         0|
#> --------------------------------------------------------
#> 
#> 
#> $IEADB
#> --------------------------------------------------------
#> |  Variable   |  Class  |  Obs  |  Missing  |  Miss \%  |
#> --------------------------------------------------------
#> |manyID       |character|   3667|          0|         0|
#> |treatyID     |character|   3667|          0|         0|
#> |Title        |character|   3667|          0|         0|
#> |Begin        |mdate    |   3667|          0|         0|
#> |Ambit        |character|   3667|          0|         0|
#> |AgreementType|character|   3667|          0|         0|
#> |Signature    |mdate    |   3602|         65|      1.77|
#> |Force        |mdate    |   2799|        868|     23.67|
#> |Lineage      |character|   3667|          0|         0|
#> |mitchID      |character|   3667|          0|         0|
#> --------------------------------------------------------
#> 
#> 
#> $HEIDI
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |manyID      |character|   2280|          0|         0|
#> |treatyID    |character|   2280|          0|         0|
#> |Title       |character|   2280|          0|         0|
#> |Begin       |mdate    |   2253|         27|      1.18|
#> |Signature   |mdate    |   2253|         27|      1.18|
#> |Lineage     |character|   2280|          0|         0|
#> |heidiID     |numeric  |   2280|          0|         0|
#> -------------------------------------------------------
#> 
#> 
#> $TFDD
#> ------------------------------------------------------------------
#> |       Variable        |  Class  |  Obs  |  Missing  |  Miss \%  |
#> ------------------------------------------------------------------
#> |manyID                 |character|    784|          0|         0|
#> |Title                  |character|    784|          0|         0|
#> |Begin                  |mdate    |    779|          5|      0.64|
#> |Signature              |mdate    |    779|          5|      0.64|
#> |AgreementType          |character|    584|        200|     25.51|
#> |Ambit                  |character|    770|         14|      1.79|
#> |GeogArea               |character|    545|        239|     30.48|
#> |Water_as_consumable_rsc|character|    783|          1|      0.13|
#> |NotTreaty              |character|    783|          1|      0.13|
#> |Issue                  |character|    550|        234|     29.85|
#> |Basin                  |character|    784|          0|         0|
#> |Lineage                |character|    439|        345|     44.01|
#> |tfddID                 |character|    784|          0|         0|
#> |treatyID               |character|    784|          0|         0|
#> ------------------------------------------------------------------
#> 
#> 
#> $TREND
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |manyID      |character|    710|          0|         0|
#> |treatyID    |character|    710|          0|         0|
#> |Title       |character|    710|          0|         0|
#> |Begin       |mdate    |    705|          5|       0.7|
#> |Signature   |mdate    |    705|          5|       0.7|
#> |Force       |mdate    |    705|          5|       0.7|
#> |trendID     |character|    710|          0|         0|
#> -------------------------------------------------------
#> 
#> 
#> $LABPTA
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |manyID      |character|    483|          0|         0|
#> |treatyID    |character|    483|          0|         0|
#> |Title       |character|    483|          0|         0|
#> |Begin       |mdate    |    483|          0|         0|
#> |Signature   |mdate    |    483|          0|         0|
#> |Force       |mdate    |    483|          0|         0|
#> |labptaID    |character|    483|          0|         0|
#> -------------------------------------------------------
#> 
#> 
#> $TOTA
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |manyID      |character|    442|          0|         0|
#> |treatyID    |character|    442|          0|         0|
#> |Title       |character|    442|          0|         0|
#> |Begin       |mdate    |    442|          0|         0|
#> |Signature   |mdate    |    442|          0|         0|
#> |Force       |mdate    |    439|          3|      0.68|
#> |totaID      |character|    442|          0|         0|
#> -------------------------------------------------------
#> 
#> 
#> $GPTAD
#> --------------------------------------------------------
#> |  Variable   |  Class  |  Obs  |  Missing  |  Miss \%  |
#> --------------------------------------------------------
#> |manyID       |character|    340|          0|         0|
#> |treatyID     |character|    340|          0|         0|
#> |Title        |character|    340|          0|         0|
#> |Begin        |mdate    |    310|         30|      8.82|
#> |AgreementType|character|    340|          0|         0|
#> |Ambit        |character|    340|          0|         0|
#> |GeogArea     |character|    265|         75|     22.06|
#> |Signature    |mdate    |    287|         53|     15.59|
#> |Force        |mdate    |    277|         63|     18.53|
#> |gptadID      |character|    340|          0|         0|
#> --------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{Mapping}{
\tabular{llllllll}{
   \emph{manytreaties} \tab \emph{IEADB} \tab \emph{HEIDI} \tab \emph{TFDD} \tab \emph{TREND} \tab \emph{LABPTA} \tab \emph{TOTA} \tab \emph{GPTAD} \cr
   Title \tab Treaty Name \tab Name.of.the.agreement \tab DocumentName \tab Trade.Agreement \tab Name \tab name \tab Common.Name \cr
   Signature \tab Signature Date \tab signature.date \tab DateSigned \tab Year \tab year \tab date_signed \tab Date.of.Signature \cr
   Force \tab Date IEA entered into force \tab  \tab  \tab Year \tab year \tab date_into_force \tab Date.of.Entry.into.Force \cr
   mitchID \tab IEA# \tab  \tab  \tab  \tab  \tab  \tab  \cr
   heidiID \tab  \tab ID \tab  \tab  \tab  \tab  \tab  \cr
   labptaID \tab  \tab  \tab  \tab  \tab Number \tab  \tab  \cr
   tfddID \tab  \tab  \tab 2016Update ID \tab  \tab  \tab  \tab  \cr
   AgreementType \tab Agreement Type (level 2) \tab  \tab DocType \tab  \tab  \tab  \tab  \cr
   Ambit \tab Inclusion \tab  \tab NumberParties \tab  \tab  \tab  \tab Type \cr
   GeogArea \tab  \tab  \tab GeoScope \tab  \tab  \tab  \tab  \cr
   Basin \tab  \tab  \tab Basin Name \tab  \tab  \tab  \tab  \cr
   Issue \tab  \tab  \tab Issue Area \tab  \tab  \tab  \tab  \cr
   Lineage \tab  \tab  \tab PrimaryTFDDID \tab  \tab  \tab  \tab  \cr
}
}

\keyword{datasets}
