% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytreaties-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements datacube}
\format{
\describe{
\item{HUGGO: }{A dataset with 4,837
observations and 59 variables:
manyID, treatyID, Title, Begin, End, Signature, Force, TreatyText, url, Domain, AgreementType, DocType, GeogArea, gengID, ieaID, ecolexID, Parties, verified, ..., Topic, and Coder.}
\item{IEADB: }{A dataset with 3,667
observations and 10 variables:
manyID, treatyID, Title, Begin, DocType, AgreementType, Signature, Force, Lineage, and ieadbID.}
\item{HEIDI: }{A dataset with 2,280
observations and 7 variables:
manyID, treatyID, Title, Begin, Signature, Lineage, and heidiID.}
\item{TREND: }{A dataset with 710
observations and 7 variables:
manyID, treatyID, Title, Begin, Signature, Force, and trendID.}
\item{LABPTA: }{A dataset with 483
observations and 7 variables:
manyID, treatyID, Title, Begin, Signature, Force, and labptaID.}
\item{TOTA: }{A dataset with 442
observations and 7 variables:
manyID, treatyID, Title, Begin, Signature, Force, and totaID.}
\item{GPTAD: }{A dataset with 340
observations and 10 variables:
manyID, treatyID, Title, Begin, AgreementType, DocType, GeogArea, Signature, Force, and gptadID.}
\item{GHHR: }{A dataset with 149
observations and 8 variables:
manyID, treatyID, Title, Begin, Region, LegalStatus, Lineage, and ghhrID.}
\item{WHO: }{A dataset with 114
observations and 6 variables:
manyID, treatyID, Title, Begin, Organisation, and Topic.}
}
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
}
\source{
\itemize{
\item{HUGGO: }{
Hollway, James, Henrique Sposito, and Jael Tan. 2021.
International agreements for manydata.
}
\item{IEADB: }{
Mitchell, Ron B. et al. 2020.
"What we know (and could know) about international environmental agreements".
\emph{Global Environmental Politics} 20.1, pp. 103-121.
\url{https://www.iea.ulaval.ca/en/agreements}
}
\item{HEIDI: }{
Morin, Jean-Frederic, and Chantal Blouin. 2019.
"How environmental treaties contribute to global health governance".
\emph{Globalization and health} 15.1, pp. 1-8.
\url{https://www.chaire-epi.ulaval.ca/en/data/heidi}
}
\item{TREND: }{
Morin, Jean-Frederic, Andreas Dür, and Lisa Lechner. 2018.
“Mapping the trade and environment nexus: Insights from a newdataset”.
\emph{Global Environmental Politics} 18.1, pp. 122-139.
\url{http://www.chaire-epi.ulaval.ca/en/trend}
}
\item{LABPTA: }{
Raess, Damien, Andreas Dür, and D. Sari. 2018.
“Protecting labor rights in preferential trade agreements: The role of trade unions, left governments, and skilled labor”.
\emph{The Review of International Organizations} 2.13, pp. 143-162.
\doi{10.1007/s11558-018-9301-z}
}
\item{TOTA: }{
Alschner, Wolfgang, Julia Seiermann, and Dmitriy Skougarevskiy. 2017.
“Text-as-data analysis of preferential trade agreements: Mapping the PTA landscape”.
UNCTAD Research Paper No. 5.
\url{https://github.com/mappingtreaties/tota.git}.
}
\item{GPTAD: }{
World Bank Group. 2014.
\emph{Global Preferential Trade Agreement Database (GPTAD)}.
Online database. World Bank Group.
\url{https://wits.worldbank.org/gptad/library.aspx}
}
\item{GHHR: }{
B. M. Meier, O. A. Cabrera, A. Ayala, et al. 2012.
“Bridging international law and rights-based litigation: Mapping health-related rights through the development of the global health and human rights database”.
\emph{Health & Human Rights.} 14, p. 20.
\url{https://www.globalhealthrights.org/instruments/instrument-region/}
}
\item{WHO: }{
World Health Organization. 2024.
\emph{WHO MiNDbank: More Inclusiveness Needed in Disability and Development}.
Online database. World Health Organization.
\url{https://extranet.who.int/mindbank/}
}
}
}
\usage{
agreements
}
\description{
The \code{agreements} datacube is a list containing
9 datasets:
HUGGO, IEADB, HEIDI, TREND, LABPTA, TOTA, GPTAD, GHHR, and WHO.
It is a work-in-progress, so please let us know if you have any comments or suggestions.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $HUGGO
#> -------------------------------------------------------------------------------------
#> |      Column Name       |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> -------------------------------------------------------------------------------------
#> |         manyID         |  character  |            4837|     0     |       0       |
#> |        treatyID        |  character  |            4837|     0     |       0       |
#> |         Title          |  character  |            4837|     0     |       0       |
#> |         Begin          |    mdate    |            4837|    57     |     1.18      |
#> |          End           |    mdate    |            4837|   3156    |     65.25     |
#> |       Signature        |    mdate    |            4837|    57     |     1.18      |
#> |         Force          |    mdate    |            4837|   1816    |     37.54     |
#> |       TreatyText       |  character  |            4837|     0     |       0       |
#> |          url           |  character  |            4837|   2796    |     57.8      |
#> |         Domain         |  character  |            4837|     0     |       0       |
#> |     AgreementType      |  character  |            4837|   2132    |     44.08     |
#> |        DocType         |  character  |            4837|   2096    |     43.33     |
#> |        GeogArea        |  character  |            4837|   2697    |     55.76     |
#> |         gengID         |  character  |            4837|   1054    |     21.79     |
#> |         ieaID          |  character  |            4837|   4142    |     85.63     |
#> |        ecolexID        |  character  |            4837|   4360    |     90.14     |
#> |        Parties         |  character  |            4837|   2567    |     53.07     |
#> |        verified        |  character  |            4837|   4789    |     99.01     |
#> |   DocValidUntilDate    |  character  |            4837|   4803    |     99.3      |
#> |         Notes          |  character  |            4837|   4790    |     99.03     |
#> |        Download        |  character  |            4837|   3939    |     81.43     |
#> |        MEA_type        |  character  |            4837|   4142    |     85.63     |
#> |         Ambit          |  character  |            4837|   4419    |     91.36     |
#> |         Region         |  character  |            4837|   4591    |     94.91     |
#> |     subject_ecolex     |  character  |            4837|   4361    |     90.16     |
#> |      subject_iea       |  character  |            4837|   4142    |     85.63     |
#> |        Keywords        |  character  |            4837|   4431    |     91.61     |
#> |        Lineage         |  character  |            4837|   4142    |     85.63     |
#> |        Sequence        |  character  |            4837|   4143    |     85.65     |
#> |       AdoptedIn        |  character  |            4837|   4411    |     91.19     |
#> |       Languages        |  character  |            4837|   4361    |     90.16     |
#> |       Appendices       |  character  |            4837|   4802    |     99.28     |
#> |       Depository       |  character  |            4837|   4451    |     92.02     |
#> |     DepositoryURL      |  character  |            4837|   4821    |     99.67     |
#> |       Published        |  character  |            4837|   4576    |     94.6      |
#> |        Website1        |  character  |            4837|   4456    |     92.12     |
#> |        Website2        |  character  |            4837|   4813    |     99.5      |
#> |      Secretariat       |  character  |            4837|   4143    |     85.65     |
#> |     SecretariatURL     |  character  |            4837|   4143    |     85.65     |
#> |          UNEP          |  character  |            4837|   4636    |     95.84     |
#> |       Supersedes       |  character  |            4837|   4820    |     99.65     |
#> |       References       |  character  |            4837|   4802    |     99.28     |
#> |       EnabledBy        |  character  |            4837|   4730    |     97.79     |
#> |       AmendedBy        |  character  |            4837|   4746    |     98.12     |
#> |          Lit           |  character  |            4837|   4292    |     88.73     |
#> |          Data          |  character  |            4837|   4300    |     88.9      |
#> |         Coded          |  character  |            4837|   4539    |     93.84     |
#> |        Abstract        |  character  |            4837|   4811    |     99.46     |
#> |        Language        |  character  |            4837|   3488    |     72.11     |
#> |   Orig_noneng_title    |  character  |            4837|   4781    |     98.84     |
#> |         match          |  character  |            4837|   4781    |     98.84     |
#> |        Citation        |  character  |            4837|   4836    |     99.98     |
#> |         Formal         |  character  |            4837|   4608    |     95.27     |
#> |        Dataset         |  character  |            4837|   4608    |     95.27     |
#> |         Source         |  character  |            4837|   4609    |     95.29     |
#> |Health_as_primary_intent|  character  |            4837|   4608    |     95.27     |
#> |        Comments        |  character  |            4837|   4814    |     99.52     |
#> |         Topic          |  character  |            4837|   4608    |     95.27     |
#> |         Coder          |  character  |            4837|     0     |       0       |
#> -------------------------------------------------------------------------------------
#> 
#> 
#> $IEADB
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            3667|     0     |       0       |
#> |   treatyID    |  character  |            3667|     0     |       0       |
#> |     Title     |  character  |            3667|     0     |       0       |
#> |     Begin     |    mdate    |            3667|     0     |       0       |
#> |    DocType    |  character  |            3667|     0     |       0       |
#> | AgreementType |  character  |            3667|     0     |       0       |
#> |   Signature   |    mdate    |            3667|    65     |     1.77      |
#> |     Force     |    mdate    |            3667|    868    |     23.67     |
#> |    Lineage    |  character  |            3667|     0     |       0       |
#> |    ieadbID    |  character  |            3667|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HEIDI
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2280|     0     |       0       |
#> |   treatyID    |  character  |            2280|     0     |       0       |
#> |     Title     |  character  |            2280|     0     |       0       |
#> |     Begin     |    mdate    |            2280|    27     |     1.18      |
#> |   Signature   |    mdate    |            2280|    27     |     1.18      |
#> |    Lineage    |  character  |            2280|     0     |       0       |
#> |    heidiID    |   numeric   |            2280|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $TREND
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             710|     0     |       0       |
#> |   treatyID    |  character  |             710|     0     |       0       |
#> |     Title     |  character  |             710|     0     |       0       |
#> |     Begin     |    mdate    |             710|     5     |      0.7      |
#> |   Signature   |    mdate    |             710|     5     |      0.7      |
#> |     Force     |    mdate    |             710|     5     |      0.7      |
#> |    trendID    |  character  |             710|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $LABPTA
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             483|     0     |       0       |
#> |   treatyID    |  character  |             483|     0     |       0       |
#> |     Title     |  character  |             483|     0     |       0       |
#> |     Begin     |    mdate    |             483|     0     |       0       |
#> |   Signature   |    mdate    |             483|     0     |       0       |
#> |     Force     |    mdate    |             483|     0     |       0       |
#> |   labptaID    |  character  |             483|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $TOTA
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             442|     0     |       0       |
#> |   treatyID    |  character  |             442|     0     |       0       |
#> |     Title     |  character  |             442|     0     |       0       |
#> |     Begin     |    mdate    |             442|     0     |       0       |
#> |   Signature   |    mdate    |             442|     0     |       0       |
#> |     Force     |    mdate    |             442|     3     |     0.68      |
#> |    totaID     |  character  |             442|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $GPTAD
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             340|     0     |       0       |
#> |   treatyID    |  character  |             340|     0     |       0       |
#> |     Title     |  character  |             340|     0     |       0       |
#> |     Begin     |    mdate    |             340|    30     |     8.82      |
#> | AgreementType |  character  |             340|     0     |       0       |
#> |    DocType    |  character  |             340|     0     |       0       |
#> |   GeogArea    |  character  |             340|    75     |     22.06     |
#> |   Signature   |    mdate    |             340|    53     |     15.59     |
#> |     Force     |    mdate    |             340|    63     |     18.53     |
#> |    gptadID    |  character  |             340|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $GHHR
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             149|     0     |       0       |
#> |   treatyID    |  character  |             149|     0     |       0       |
#> |     Title     |  character  |             149|     0     |       0       |
#> |     Begin     |    mdate    |             149|     3     |     2.01      |
#> |    Region     |  character  |             149|     0     |       0       |
#> |  LegalStatus  |  character  |             149|     0     |       0       |
#> |    Lineage    |  character  |             149|    52     |     34.9      |
#> |    ghhrID     |  character  |             149|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $WHO
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             114|     0     |       0       |
#> |   treatyID    |  character  |             114|     0     |       0       |
#> |     Title     |  character  |             114|     0     |       0       |
#> |     Begin     |    mdate    |             114|     0     |       0       |
#> | Organisation  |  character  |             114|     0     |       0       |
#> |     Topic     |  character  |             114|     0     |       0       |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{Mapping}{
\tabular{lllllllll}{
   \emph{manytreaties} \tab \emph{IEADB} \tab \emph{HEIDI} \tab \emph{TREND} \tab \emph{LABPTA} \tab \emph{TOTA} \tab \emph{GPTAD} \tab \emph{GHHR} \tab \emph{WHO} \cr
   Title \tab Treaty Name \tab Name.of.the.agreement \tab Trade.Agreement \tab Name \tab name \tab Common.Name \tab  \tab  \cr
   Signature \tab Signature Date \tab signature.date \tab Year \tab year \tab date_signed \tab Date.of.Signature \tab  \tab  \cr
   Force \tab Date IEA entered into force \tab  \tab Year \tab year \tab date_into_force \tab Date.of.Entry.into.Force \tab Year of adoption \tab  \cr
   mitchID \tab IEA# \tab  \tab  \tab  \tab  \tab  \tab  \tab  \cr
   heidiID \tab  \tab ID \tab  \tab  \tab  \tab  \tab  \tab  \cr
   labptaID \tab  \tab  \tab  \tab Number \tab  \tab  \tab  \tab  \cr
   AgreementType \tab Agreement Type (level 2) \tab  \tab  \tab  \tab  \tab  \tab  \tab  \cr
   DocType \tab Inclusion \tab  \tab  \tab  \tab  \tab Type \tab  \tab  \cr
}
}

\keyword{datasets}
