% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytreaties-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements datacube documentation}
\format{
The agreements datacube is a list that contains the
following 8 datasets: HUGGO, IEADB, HEIDI, TREND, LABPTA, TOTA, GPTAD, TFDD.
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.

\describe{
\item{HUGGO \if{html}{\out{<img src="https://mirrors.creativecommons.org/presskit/buttons/88x31/png/by.png" height="12"/>}}:}{A dataset with 4,837
observations and 30 variables:
manyID, treatyID, Title, Begin, End, Signature, Force, Term, TreatyText, TextURL, Domain, AgreementType, DocType, Ambit, Region, Keywords, Lineage, AdoptedIn, ..., Comments, and gengID.}
\item{IEADB: }{A dataset with 3,667
observations and 10 variables:
manyID, treatyID, Title, Begin, DocType, AgreementType, Signature, Force, Lineage, and mitchID.}
\item{HEIDI: }{A dataset with 2,280
observations and 7 variables:
manyID, treatyID, Title, Begin, Signature, Lineage, and heidiID.}
\item{TFDD: }{A dataset with 784
observations and 14 variables:
manyID, Title, Begin, Signature, AgreementType, DocType, GeogArea, Water_as_consumable_rsc, NotTreaty, Issue, Basin, Lineage, tfddID, and treatyID.}
\item{TREND: }{A dataset with 710
observations and 7 variables:
manyID, treatyID, Title, Begin, Signature, Force, and trendID.}
\item{LABPTA \if{html}{\out{<img src="https://mirrors.creativecommons.org/presskit/buttons/88x31/png/by.png" height="12"/>}}:}{A dataset with 483
observations and 7 variables:
manyID, treatyID, Title, Begin, Signature, Force, and labptaID.}
\item{TOTA \if{html}{\out{<img src="https://mirrors.creativecommons.org/presskit/buttons/88x31/png/by.png" height="12"/>}}:}{A dataset with 442
observations and 7 variables:
manyID, treatyID, Title, Begin, Signature, Force, and totaID.}
\item{GPTAD \if{html}{\out{<img src="https://mirrors.creativecommons.org/presskit/buttons/88x31/png/by.png" height="12"/>}}:}{A dataset with 340
observations and 10 variables:
manyID, treatyID, Title, Begin, AgreementType, DocType, GeogArea, Signature, Force, and gptadID.}
}
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
}
\source{
\itemize{
\item{HUGGO: }{
Hollway, James, Henrique Sposito, and Jael Tan. 2021.
International agreements for manydata.
}
\item{IEADB: }{
Mitchell, Ron B. et al. 2020.
"What we know (and could know) about international environmental agreements".
\emph{Global Environmental Politics} 20.1, pp. 103-121.
\url{https://www.iea.ulaval.ca/en/agreements}
}
\item{HEIDI: }{
Morin, Jean-Frederic, and Chantal Blouin. 2019.
"How environmental treaties contribute to global health governance".
\emph{Globalization and health} 15.1, pp. 1-8.
\url{https://www.chaire-epi.ulaval.ca/en/data/heidi}
}
\item{TFDD: }{
College of Earth, Ocean and Atmospheric Science, Oregon State University. 2021.
Product of the Transboundary FreshwaterDispute Database.
\url{http://transboundarywaters.science.oregonstate.edu}
}
\item{TREND: }{
Morin, Jean-Frederic, Andreas Dür, and Lisa Lechner. 2018.
"Mapping the trade and environment nexus: Insights from a new dataset".
\emph{Global Environmental Politics} 18.1, pp. 122-139.
\url{http://www.chaire-epi.ulaval.ca/en/trend}
}
\item{LABPTA: }{
Raess, Damien, Andreas Dür, and D. Sari. 2018.
"Protecting labor rights in preferential trade agreements: The role of trade unions, left governments, and skilled labor".
\emph{The Review of International Organizations} 2.13, pp. 143-162.
\doi{10.1007/s11558-018-9301-z}
}
\item{TOTA: }{
Alschner, Wolfgang, Julia Seiermann, and Dmitriy Skougarevskiy. 2017.
"Text-as-data analysis of preferential trade agreements: Mapping the PTA landscape".
UNCTAD Research Paper No. 5.
\url{https://github.com/mappingtreaties/tota.git}.
}
\item{GPTAD: }{
World Bank Group. 2014.
"Global Preferential Trade Agreement Database (GPTAD)".
Online database. \emph{World Bank Group}.
\url{https://wits.worldbank.org/gptad/library.aspx}
}
}
}
\usage{
agreements
}
\description{
agreements datacube documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $HUGGO
#> --------------------------------------------------------
#> |  Variable   |  Class  |  Obs  |  Missing  |  Miss \%  |
#> --------------------------------------------------------
#> |manyID       |character|   4837|          0|         0|
#> |treatyID     |character|   4837|          0|         0|
#> |Title        |character|   4837|          0|         0|
#> |Begin        |mdate    |   4780|         57|      1.18|
#> |End          |mdate    |   1681|       3156|     65.25|
#> |Signature    |mdate    |   4780|         57|      1.18|
#> |Force        |mdate    |   3021|       1816|     37.54|
#> |Term         |mdate    |     32|       4805|     99.34|
#> |TreatyText   |logical  |   4837|          0|         0|
#> |TextURL      |character|   2043|       2794|     57.76|
#> |Domain       |character|   4758|         79|      1.63|
#> |Topic        |character|    229|       4608|     95.27|
#> |AgreementType|character|   2705|       2132|     44.08|
#> |MEA_type     |character|    695|       4142|     85.63|
#> |DocType      |character|   2741|       2096|     43.33|
#> |Ambit        |character|   2335|       2502|     51.73|
#> |Region       |character|    246|       4591|     94.91|
#> |Keywords     |character|    406|       4431|     91.61|
#> |Lineage      |character|    695|       4142|     85.63|
#> |AdoptedIn    |character|    426|       4411|     91.19|
#> |Depository   |character|    386|       4451|     92.02|
#> |Published    |character|    261|       4576|      94.6|
#> |Website      |character|    843|       3994|     82.57|
#> |Secretariat  |character|    694|       4143|     85.65|
#> |Abstract     |character|     26|       4811|     99.46|
#> |Appendices   |numeric  |     35|       4802|     99.28|
#> |Language     |character|   1349|       3488|     72.11|
#> |Languages    |character|    476|       4361|     90.16|
#> |OrigTitle    |character|     56|       4781|     98.84|
#> |match        |character|     56|       4781|     98.84|
#> |Coder        |character|   4837|          0|         0|
#> |Comments     |character|    358|       4479|      92.6|
#> |gengID       |character|   3783|       1054|     21.79|
#> --------------------------------------------------------
#> 
#> 
#> $IEADB
#> --------------------------------------------------------
#> |  Variable   |  Class  |  Obs  |  Missing  |  Miss \%  |
#> --------------------------------------------------------
#> |manyID       |character|   3667|          0|         0|
#> |treatyID     |character|   3667|          0|         0|
#> |Title        |character|   3667|          0|         0|
#> |Begin        |mdate    |   3667|          0|         0|
#> |DocType      |character|   3667|          0|         0|
#> |AgreementType|character|   3667|          0|         0|
#> |Signature    |mdate    |   3602|         65|      1.77|
#> |Force        |mdate    |   2799|        868|     23.67|
#> |Lineage      |character|   3667|          0|         0|
#> |mitchID      |character|   3667|          0|         0|
#> --------------------------------------------------------
#> 
#> 
#> $HEIDI
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |manyID      |character|   2280|          0|         0|
#> |treatyID    |character|   2280|          0|         0|
#> |Title       |character|   2280|          0|         0|
#> |Begin       |mdate    |   2253|         27|      1.18|
#> |Signature   |mdate    |   2253|         27|      1.18|
#> |Lineage     |character|   2280|          0|         0|
#> |heidiID     |numeric  |   2280|          0|         0|
#> -------------------------------------------------------
#> 
#> 
#> $TFDD
#> ------------------------------------------------------------------
#> |       Variable        |  Class  |  Obs  |  Missing  |  Miss \%  |
#> ------------------------------------------------------------------
#> |manyID                 |character|    784|          0|         0|
#> |Title                  |character|    784|          0|         0|
#> |Begin                  |mdate    |    779|          5|      0.64|
#> |Signature              |mdate    |    779|          5|      0.64|
#> |AgreementType          |character|    584|        200|     25.51|
#> |DocType                |character|    770|         14|      1.79|
#> |GeogArea               |character|    545|        239|     30.48|
#> |Water_as_consumable_rsc|character|    783|          1|      0.13|
#> |NotTreaty              |character|    783|          1|      0.13|
#> |Issue                  |character|    550|        234|     29.85|
#> |Basin                  |character|    784|          0|         0|
#> |Lineage                |character|    439|        345|     44.01|
#> |tfddID                 |character|    784|          0|         0|
#> |treatyID               |character|    784|          0|         0|
#> ------------------------------------------------------------------
#> 
#> 
#> $TREND
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |manyID      |character|    710|          0|         0|
#> |treatyID    |character|    710|          0|         0|
#> |Title       |character|    710|          0|         0|
#> |Begin       |mdate    |    705|          5|       0.7|
#> |Signature   |mdate    |    705|          5|       0.7|
#> |Force       |mdate    |    705|          5|       0.7|
#> |trendID     |character|    710|          0|         0|
#> -------------------------------------------------------
#> 
#> 
#> $LABPTA
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |manyID      |character|    483|          0|         0|
#> |treatyID    |character|    483|          0|         0|
#> |Title       |character|    483|          0|         0|
#> |Begin       |mdate    |    483|          0|         0|
#> |Signature   |mdate    |    483|          0|         0|
#> |Force       |mdate    |    483|          0|         0|
#> |labptaID    |character|    483|          0|         0|
#> -------------------------------------------------------
#> 
#> 
#> $TOTA
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |manyID      |character|    442|          0|         0|
#> |treatyID    |character|    442|          0|         0|
#> |Title       |character|    442|          0|         0|
#> |Begin       |mdate    |    442|          0|         0|
#> |Signature   |mdate    |    442|          0|         0|
#> |Force       |mdate    |    439|          3|      0.68|
#> |totaID      |character|    442|          0|         0|
#> -------------------------------------------------------
#> 
#> 
#> $GPTAD
#> --------------------------------------------------------
#> |  Variable   |  Class  |  Obs  |  Missing  |  Miss \%  |
#> --------------------------------------------------------
#> |manyID       |character|    340|          0|         0|
#> |treatyID     |character|    340|          0|         0|
#> |Title        |character|    340|          0|         0|
#> |Begin        |mdate    |    310|         30|      8.82|
#> |AgreementType|character|    340|          0|         0|
#> |DocType      |character|    340|          0|         0|
#> |GeogArea     |character|    265|         75|     22.06|
#> |Signature    |mdate    |    287|         53|     15.59|
#> |Force        |mdate    |    277|         63|     18.53|
#> |gptadID      |character|    340|          0|         0|
#> --------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{Mapping}{
\tabular{llllllll}{
   \emph{manytreaties} \tab \emph{IEADB} \tab \emph{HEIDI} \tab \emph{TFDD} \tab \emph{TREND} \tab \emph{LABPTA} \tab \emph{TOTA} \tab \emph{GPTAD} \cr
   Title \tab Treaty Name \tab Name.of.the.agreement \tab DocumentName \tab Trade.Agreement \tab Name \tab name \tab Common.Name \cr
   Signature \tab Signature Date \tab signature.date \tab DateSigned \tab Year \tab year \tab date_signed \tab Date.of.Signature \cr
   Force \tab Date IEA entered into force \tab  \tab  \tab Year \tab year \tab date_into_force \tab Date.of.Entry.into.Force \cr
   mitchID \tab IEA# \tab  \tab  \tab  \tab  \tab  \tab  \cr
   heidiID \tab  \tab ID \tab  \tab  \tab  \tab  \tab  \cr
   labptaID \tab  \tab  \tab  \tab  \tab Number \tab  \tab  \cr
   tfddID \tab  \tab  \tab 2016Update ID \tab  \tab  \tab  \tab  \cr
   AgreementType \tab Agreement Type (level 2) \tab  \tab DocType \tab  \tab  \tab  \tab  \cr
   DocType \tab Inclusion \tab  \tab NumberParties \tab  \tab  \tab  \tab Type \cr
   GeogArea \tab  \tab  \tab GeoScope \tab  \tab  \tab  \tab  \cr
   Basin \tab  \tab  \tab Basin Name \tab  \tab  \tab  \tab  \cr
   Issue \tab  \tab  \tab Issue Area \tab  \tab  \tab  \tab  \cr
   Lineage \tab  \tab  \tab PrimaryTFDDID \tab  \tab  \tab  \tab  \cr
}
}

\keyword{datasets}
