% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytreaties-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships datacube documentation}
\format{
The memberships datacube is a list that contains the
following 4 datasets: GPTAD_MEM, IEADB_MEM, TFDD_MEM, HUGGO_MEM.
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
\describe{
\item{GPTAD_MEM: }{A dataset with 2192 observations and the following
9 variables: manyID, treatyID, Title, Begin, stateID, Signature, Force,
StateName, gptadID.}
\item{IEADB_MEM: }{A dataset with 15463 observations and the following
12 variables: manyID, treatyID, Title, Begin, stateID, End, stateSignature,
Signature, Rat, Force, DocType, ieadbID.}
\item{TFDD_MEM: }{A dataset with 1832 observations and the following
8 variables: manyID, treatyID, Title, Begin, stateID, Signature, tfddID,
Memberships.}
\item{HUGGO_MEM: }{A dataset with 91729 observations and the following
41 variables: manyID, treatyID, Title, Begin, stateID, StateName,
StateSignature, StateRatification, StateForce, StateEnd, Rat=Notif,
Accession, Succession, Signature, Force, End, gengID, url, ieaID, ecolexID,
StateForce2, StateForce3, StateEnd2, Term, Comments, Deposit, obsolete,
ProvisionalApp, Reservation, Notes, stateForce_ecolex, stateForce_iea,
Consent, Acceptance, Orig_noneng_title, match, gptadID, destaID, Title2,
Domain, Coder.}
}
}
\source{
\itemize{
\item{GPTAD_MEM: }{
W. B. Group.
\emph{Global Preferential Trade Agreement Database (GPTAD)}.
\url{https://wits.worldbank.org/gptad/library.aspx}. 2013.}
\item{IEADB_MEM: }{
R. B. Mitchell, L. B. Andonova, M. Axelrod, et al. “What we know (and could know) about international environmental agreements”.
\emph{Global Environmental Politics} 20.1 (2020), pp. 103-121.}
\item{TFDD_MEM: }{
O. College of Earth and O. S. U. Atmospheric Science.
\emph{Product of the Transboundary Freshwater DisputeDatabase}.
\url{http://transboundarywaters.science.oregonstate.edu.}. 2021.}
}
}
\usage{
memberships
}
\description{
memberships datacube documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $GPTAD_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2192|     0     |       0       |
#> |   treatyID    |  character  |            2192|     0     |       0       |
#> |     Title     |  character  |            2192|     0     |       0       |
#> |     Begin     |    mdate    |            2192|    99     |     4.52      |
#> |    stateID    |  character  |            2192|    55     |     2.51      |
#> |   Signature   |    mdate    |            2192|    310    |     14.14     |
#> |     Force     |    mdate    |            2192|    285    |      13       |
#> |   StateName   |  character  |            2192|    36     |     1.64      |
#> |    gptadID    |  character  |            2192|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $IEADB_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |           15463|     0     |       0       |
#> |   treatyID    |  character  |           15463|     0     |       0       |
#> |     Title     |  character  |           15463|     0     |       0       |
#> |     Begin     |    mdate    |           15463|     0     |       0       |
#> |    stateID    |  character  |           15463|     0     |       0       |
#> |      End      |    mdate    |           15463|   14593   |     94.37     |
#> |stateSignature |    mdate    |           15463|   5797    |     37.49     |
#> |   Signature   |    mdate    |           15463|     0     |       0       |
#> |      Rat      |    mdate    |           15463|   5485    |     35.47     |
#> |     Force     |    mdate    |           15463|     0     |       0       |
#> |    DocType    |  character  |           15463|     0     |       0       |
#> |    ieadbID    |   numeric   |           15463|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $TFDD_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            1832|     0     |       0       |
#> |   treatyID    |  character  |            1832|     0     |       0       |
#> |     Title     |  character  |            1832|     1     |     0.05      |
#> |     Begin     |    mdate    |            1832|    102    |     5.57      |
#> |    stateID    |  character  |            1832|     1     |     0.05      |
#> |   Signature   |    mdate    |            1832|    102    |     5.57      |
#> |    tfddID     |   numeric   |            1832|     1     |     0.05      |
#> |  Memberships  |  character  |            1832|    805    |     43.94     |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO_MEM
#> ------------------------------------------------------------------------------
#> |   Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ------------------------------------------------------------------------------
#> |     manyID      |  character  |           91729|     0     |       0       |
#> |      Title      |  character  |           91729|     0     |       0       |
#> |      Begin      |    mdate    |           91729|    12     |     0.01      |
#> |     stateID     |  character  |           91729|     0     |       0       |
#> |    StateName    |  character  |           91729|   50485   |     55.04     |
#> | StateSignature  |    mdate    |           91729|   17944   |     19.56     |
#> |StateRatification|    mdate    |           91729|   63960   |     69.73     |
#> |   StateForce    |    mdate    |           91729|   50851   |     55.44     |
#> |    StateEnd     |    mdate    |           91729|   33471   |     36.49     |
#> |    Rat=Notif    |  character  |           91729|   90502   |     98.66     |
#> |    Accession    |    mdate    |           91729|   77230   |     84.19     |
#> |   Succession    |  character  |           91729|     0     |       0       |
#> |    treatyID     |  character  |           91729|     0     |       0       |
#> |    Signature    |    mdate    |           91729|    16     |     0.02      |
#> |      Force      |    mdate    |           91729|   34549   |     37.66     |
#> |       End       |    mdate    |           91729|   38792   |     42.29     |
#> |     gengID      |  character  |           91729|   40770   |     44.45     |
#> |       url       |  character  |           91729|   90798   |     98.99     |
#> |      ieaID      |  character  |           91729|   49378   |     53.83     |
#> |    ecolexID     |  character  |           91729|   71730   |     78.2      |
#> |   StateForce2   |    mdate    |           91729|   85042   |     92.71     |
#> |   StateForce3   |    mdate    |           91729|   88039   |     95.98     |
#> |    StateEnd2    |    mdate    |           91729|   88033   |     95.97     |
#> |      Term       |    mdate    |           91729|   90551   |     98.72     |
#> |    Comments     |  character  |           91729|   88449   |     96.42     |
#> |     Deposit     |  character  |           91729|   90079   |     98.2      |
#> |    obsolete     |  character  |           91729|   79154   |     86.29     |
#> | ProvisionalApp  |  character  |           91729|   90957   |     99.16     |
#> |   Reservation   |  character  |           91729|   90973   |     99.18     |
#> |      Notes      |  character  |           91729|   91543   |     99.8      |
#> |stateForce_ecolex|    mdate    |           91729|   73212   |     79.81     |
#> | stateForce_iea  |    mdate    |           91729|   73212   |     79.81     |
#> |     Consent     |  character  |           91729|   91505   |     99.76     |
#> |   Acceptance    |  character  |           91729|   91204   |     99.43     |
#> |Orig_noneng_title|  character  |           91729|   91725   |      100      |
#> |      match      |  character  |           91729|   91725   |      100      |
#> |     gptadID     |  character  |           91729|   89560   |     97.64     |
#> |     destaID     |  character  |           91729|   86694   |     94.51     |
#> |     Title2      |  character  |           91729|   91252   |     99.48     |
#> |     Domain      |  character  |           91729|     0     |       0       |
#> |      Coder      |  character  |           91729|    15     |     0.02      |
#> ------------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{URL}{

\itemize{
\item{GPTAD_MEM: }{
\url{https://wits.worldbank.org/gptad/library.aspx}
}
\item{IEADB_MEM: }{
\url{https://iea.uoregon.edu/country-members}
}
\item{TFDD_MEM: }{
\url{https://transboundarywaters.science.oregonstate.edu/}
}
\item{HUGGO_MEM: }{Hand-coded data by the GGO team.}
}
}

\section{Mapping}{

\itemize{
\item{GPTAD_MEM: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   Membership \tab StateName \cr
   Common.Name \tab Title \cr
   Date.of.Signature \tab Signature \cr
   Date.of.Entry.into.Force \tab Force \cr
}


}
\item{IEADB_MEM: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   country \tab stateID \cr
   treatyname \tab Title \cr
   tsig \tab Signature \cr
   csig \tab stateSignature \cr
   crat \tab Rat \cr
   tterm \tab End \cr
   ceif3 \tab Force \cr
   ceif4 \tab Force2 \cr
   mitch_id \tab ieadbID \cr
   inclusion \tab DocType \cr
}


}
\item{TFDD_MEM: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   DateSigned \tab Signature \cr
   '2016Update ID' \tab tfddID \cr
   CCODE \tab stateID \cr
   DocumentName \tab Title \cr
   Signatories \tab Memberships \cr
}


}
\item{HUGGO_MEM: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   NA \tab NA \cr
}


}
}
}

\keyword{datasets}
