% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytreaties-parties.R
\docType{data}
\name{parties}
\alias{parties}
\title{parties datacube}
\format{
\describe{
\item{HUGGO \if{html}{\out{<img src="https://mirrors.creativecommons.org/presskit/buttons/88x31/png/by.png" height="12"/>}}:}{A dataset with
observations and  variables:
.}
\item{IEADB: }{A dataset with 15,463
observations and 10 variables:
stateID, treatyID, Begin, End, Title, stateSignature, Signature, Rat, Force, and ieadbID.}
\item{DESTA: }{A dataset with 7,466
observations and 8 variables:
stateID, treatyID, Begin, Title, Signature, Force, StateName, and destaID.}
\item{GPTAD: }{A dataset with 2,192
observations and 6 variables:
stateID, treatyID, Begin, Title, Signature, and Force.}
\item{TFDD: }{A dataset with 1,832
observations and 6 variables:
stateID, treatyID, Begin, Title, Signature, and tfddID.}
}
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
}
\source{
\itemize{
\item Mitchell, Ron B. et al. 2020. 'What we know (and could know) about international environmental agreements'. \emph{Global Environmental Politics} 20.1, pp. 103-121.
\item Duer, Andreas, Leonardo Baccini, and Manfred Elsig. 2014. 'The Design of International Trade Agreements: Introducing a New Database'. \emph{The Review of International Organizations} 9.3, pp. 353-375.
\item World Bank Group. 2014. \emph{Global Preferential Trade Agreement Database (GPTAD)}. Online database. World Bank Group.
\item Oregon College of Earth and Oregon State University Atmospheric Science. 2021. \emph{Product of the Transboundary Freshwater Dispute Database}.
\item Hollway, James, and Jael Tan. 'Global governance observations on parties to international agreements' (2025).
}
}
\usage{
parties
}
\description{
The \code{manytreaties::parties} datacube is a list containing 5 datasets: IEADB, DESTA, GPTAD, TFDD, and GGO.
It is a work-in-progress, so please let us know if you have any comments or suggestions.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $IEADB
#> ---------------------------------------------------------
#> |   Variable   |  Class  |  Obs  |  Missing  |  Miss \%  |
#> ---------------------------------------------------------
#> |stateID       |character|  15463|          0|         0|
#> |treatyID      |character|  15463|          0|         0|
#> |Begin         |mdate    |  15463|          0|         0|
#> |End           |mdate    |    870|      14593|     94.37|
#> |Title         |character|  15463|          0|         0|
#> |stateSignature|mdate    |   9666|       5797|     37.49|
#> |Signature     |mdate    |  15463|          0|         0|
#> |Rat           |mdate    |   9978|       5485|     35.47|
#> |Force         |mdate    |  15463|          0|         0|
#> |ieadbID       |numeric  |  15463|          0|         0|
#> ---------------------------------------------------------
#> 
#> 
#> $DESTA
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |stateID     |character|   7455|         11|      0.15|
#> |treatyID    |character|   7466|          0|         0|
#> |Begin       |mdate    |   6836|        630|      8.44|
#> |Title       |character|   7466|          0|         0|
#> |Signature   |mdate    |   6836|        630|      8.44|
#> |Force       |mdate    |   4602|       2864|     38.36|
#> |StateName   |character|   7466|          0|         0|
#> |destaID     |character|   7466|          0|         0|
#> -------------------------------------------------------
#> 
#> 
#> $GPTAD
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |stateID     |character|   2137|         55|      2.51|
#> |treatyID    |character|   2192|          0|         0|
#> |Begin       |mdate    |   2093|         99|      4.52|
#> |Title       |character|   2192|          0|         0|
#> |Signature   |mdate    |   1882|        310|     14.14|
#> |Force       |mdate    |   1907|        285|        13|
#> -------------------------------------------------------
#> 
#> 
#> $TFDD
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |stateID     |character|   1831|          1|      0.05|
#> |treatyID    |character|   1832|          0|         0|
#> |Begin       |mdate    |   1730|        102|      5.57|
#> |Title       |character|   1831|          1|      0.05|
#> |Signature   |mdate    |   1730|        102|      5.57|
#> |tfddID      |numeric  |   1831|          1|      0.05|
#> -------------------------------------------------------
#> 
#> 
#> $GGO
#> ------------------------------------------------------------
#> |    Variable     |  Class  |  Obs  |  Missing  |  Miss \%  |
#> ------------------------------------------------------------
#> |stateID          |character|  98320|          0|         0|
#> |treatyID         |character|  98320|          0|         0|
#> |Title            |character|  98320|          0|         0|
#> |Begin            |mdate    |  98320|          0|         0|
#> |End              |mdate    |  68402|      29918|     30.43|
#> |Signature        |mdate    |  98320|          0|         0|
#> |StateSignature   |mdate    |  80137|      18183|     18.49|
#> |StateRatification|mdate    |  29013|      69307|     70.49|
#> |Force            |mdate    |  63700|      34620|     35.21|
#> |StateForce       |mdate    |  62574|      35746|     36.36|
#> |Term             |Date     |   1167|      97153|     98.81|
#> |StateEnd         |mdate    |  61895|      36425|     37.05|
#> |Coder            |character|  98305|         15|      0.02|
#> |Comments         |character|   4642|      93678|     95.28|
#> ------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{Mapping}{
\tabular{lllll}{
   \emph{manytreaties} \tab \emph{IEADB} \tab \emph{DESTA} \tab \emph{GPTAD} \tab \emph{TFDD} \cr
   stateID \tab country \tab Member \tab  \tab CCODE \cr
   Signature \tab csig \tab year \tab Date.of.Signature \tab DateSigned \cr
   Rat \tab crat \tab  \tab  \tab  \cr
   Force \tab ceif3 \tab entryforceyear \tab Date.of.Entry.into.Force \tab  \cr
   Force2 \tab ceif4 \tab  \tab  \tab  \cr
   End \tab tterm \tab  \tab  \tab  \cr
   Title \tab treatyname \tab name \tab Common.Name \tab DocumentName \cr
   Memberships \tab  \tab  \tab StateName \tab Signatories \cr
   destaID \tab  \tab base_treaty \tab  \tab  \cr
   tfddID \tab  \tab  \tab  \tab 2016Update ID \cr
   mitchID \tab ieadbID \tab  \tab  \tab  \cr
   DocType \tab inclusion \tab  \tab  \tab  \cr
}
}

\keyword{datasets}
