% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytreaties-parties.R
\docType{data}
\name{parties}
\alias{parties}
\title{parties datacube}
\format{
\describe{
\item{HUGGO \if{html}{\out{<img src="https://mirrors.creativecommons.org/presskit/buttons/88x31/png/by.png" height="12"/>}}:}{A dataset with 91,729
observations and 41 variables:
manyID, Title, Begin, stateID, StateName, StateSignature, StateRatification, StateForce, StateEnd, Rat=Notif, Accession, Succession, treatyID, Signature, Force, End, gengID, url, ..., Domain, and Coder.}
\item{IEADB: }{A dataset with 15,463
observations and 12 variables:
manyID, treatyID, Title, Begin, stateID, End, stateSignature, Signature, Rat, Force, DocType, and ieadbID.}
\item{GPTAD \if{html}{\out{<img src="https://mirrors.creativecommons.org/presskit/buttons/88x31/png/by.png" height="12"/>}}:}{A dataset with 2,192
observations and 9 variables:
manyID, treatyID, Title, Begin, stateID, Signature, Force, StateName, and gptadID.}
\item{TFDD \if{html}{\out{<img src="https://mirrors.creativecommons.org/presskit/buttons/88x31/png/cc-zero.png" height="12"/>}}:}{A dataset with 1,832
observations and 8 variables:
manyID, treatyID, Title, Begin, stateID, Signature, tfddID, and Memberships.}
}
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
}
\source{
\itemize{
\item{HUGGO: }{
Hollway, James, Henrique Sposito, and Jael Tan. 2021.
International agreements for manydata.
}
\item{IEADB: }{
Mitchell, Ron B. et al. 2020.
"What we know (and could know) about international environmental agreements".
\emph{Global Environmental Politics} 20.1, pp. 103-121.
\url{https://www.iea.ulaval.ca/en/country-members}
}
\item{GPTAD: }{
World Bank Group. 2014.
\emph{Global Preferential Trade Agreement Database (GPTAD)}.
Online database. World Bank Group.
\url{https://wits.worldbank.org/gptad/library.aspx}
}
\item{TFDD: }{
Oregon College of Earth and Oregon State University Atmospheric Science. 2021.
\emph{Product of the Transboundary Freshwater Dispute Database}.
\url{http://transboundarywaters.science.oregonstate.edu}.
}
}
}
\usage{
parties
}
\description{
The \code{parties} datacube is a list containing
4 datasets:
HUGGO, IEADB, GPTAD, and TFDD.
It is a work-in-progress, so please let us know if you have any comments or suggestions.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $HUGGO
#> ------------------------------------------------------------------------------
#> |   Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ------------------------------------------------------------------------------
#> |     manyID      |  character  |           91729|     0     |       0       |
#> |      Title      |  character  |           91729|     0     |       0       |
#> |      Begin      |    mdate    |           91729|    12     |     0.01      |
#> |     stateID     |  character  |           91729|     0     |       0       |
#> |    StateName    |  character  |           91729|   50485   |     55.04     |
#> | StateSignature  |    mdate    |           91729|   17944   |     19.56     |
#> |StateRatification|    mdate    |           91729|   63960   |     69.73     |
#> |   StateForce    |    mdate    |           91729|   50851   |     55.44     |
#> |    StateEnd     |    mdate    |           91729|   33471   |     36.49     |
#> |    Rat=Notif    |  character  |           91729|   90502   |     98.66     |
#> |    Accession    |    mdate    |           91729|   77230   |     84.19     |
#> |   Succession    |  character  |           91729|     0     |       0       |
#> |    treatyID     |  character  |           91729|     0     |       0       |
#> |    Signature    |    mdate    |           91729|    16     |     0.02      |
#> |      Force      |    mdate    |           91729|   34549   |     37.66     |
#> |       End       |    mdate    |           91729|   38792   |     42.29     |
#> |     gengID      |  character  |           91729|   40770   |     44.45     |
#> |       url       |  character  |           91729|   90798   |     98.99     |
#> |      ieaID      |  character  |           91729|   49378   |     53.83     |
#> |    ecolexID     |  character  |           91729|   71730   |     78.2      |
#> |   StateForce2   |    mdate    |           91729|   85042   |     92.71     |
#> |   StateForce3   |    mdate    |           91729|   88039   |     95.98     |
#> |    StateEnd2    |    mdate    |           91729|   88033   |     95.97     |
#> |      Term       |    mdate    |           91729|   90551   |     98.72     |
#> |    Comments     |  character  |           91729|   88449   |     96.42     |
#> |     Deposit     |  character  |           91729|   90079   |     98.2      |
#> |    obsolete     |  character  |           91729|   79154   |     86.29     |
#> | ProvisionalApp  |  character  |           91729|   90957   |     99.16     |
#> |   Reservation   |  character  |           91729|   90973   |     99.18     |
#> |      Notes      |  character  |           91729|   91543   |     99.8      |
#> |stateForce_ecolex|    mdate    |           91729|   73212   |     79.81     |
#> | stateForce_iea  |    mdate    |           91729|   73212   |     79.81     |
#> |     Consent     |  character  |           91729|   91505   |     99.76     |
#> |   Acceptance    |  character  |           91729|   91204   |     99.43     |
#> |Orig_noneng_title|  character  |           91729|   91725   |      100      |
#> |      match      |  character  |           91729|   91725   |      100      |
#> |     gptadID     |  character  |           91729|   89560   |     97.64     |
#> |     destaID     |  character  |           91729|   86694   |     94.51     |
#> |     Title2      |  character  |           91729|   91252   |     99.48     |
#> |     Domain      |  character  |           91729|     0     |       0       |
#> |      Coder      |  character  |           91729|    15     |     0.02      |
#> ------------------------------------------------------------------------------
#> 
#> 
#> $IEADB
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |           15463|     0     |       0       |
#> |   treatyID    |  character  |           15463|     0     |       0       |
#> |     Title     |  character  |           15463|     0     |       0       |
#> |     Begin     |    mdate    |           15463|     0     |       0       |
#> |    stateID    |  character  |           15463|     0     |       0       |
#> |      End      |    mdate    |           15463|   14593   |     94.37     |
#> |stateSignature |    mdate    |           15463|   5797    |     37.49     |
#> |   Signature   |    mdate    |           15463|     0     |       0       |
#> |      Rat      |    mdate    |           15463|   5485    |     35.47     |
#> |     Force     |    mdate    |           15463|     0     |       0       |
#> |    DocType    |  character  |           15463|     0     |       0       |
#> |    ieadbID    |   numeric   |           15463|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $GPTAD
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2192|     0     |       0       |
#> |   treatyID    |  character  |            2192|     0     |       0       |
#> |     Title     |  character  |            2192|     0     |       0       |
#> |     Begin     |    mdate    |            2192|    99     |     4.52      |
#> |    stateID    |  character  |            2192|    55     |     2.51      |
#> |   Signature   |    mdate    |            2192|    310    |     14.14     |
#> |     Force     |    mdate    |            2192|    285    |      13       |
#> |   StateName   |  character  |            2192|    36     |     1.64      |
#> |    gptadID    |  character  |            2192|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $TFDD
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            1832|     0     |       0       |
#> |   treatyID    |  character  |            1832|     0     |       0       |
#> |     Title     |  character  |            1832|     1     |     0.05      |
#> |     Begin     |    mdate    |            1832|    102    |     5.57      |
#> |    stateID    |  character  |            1832|     1     |     0.05      |
#> |   Signature   |    mdate    |            1832|    102    |     5.57      |
#> |    tfddID     |   numeric   |            1832|     1     |     0.05      |
#> |  Memberships  |  character  |            1832|    805    |     43.94     |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{Mapping}{
\tabular{llll}{
   \emph{manytreaties} \tab \emph{IEADB} \tab \emph{GPTAD} \tab \emph{TFDD} \cr
   stateID \tab country \tab  \tab CCODE \cr
   Signature \tab csig \tab Date.of.Signature \tab DateSigned \cr
   Rat \tab crat \tab  \tab  \cr
   Force \tab ceif3 \tab Date.of.Entry.into.Force \tab  \cr
   Force2 \tab ceif4 \tab  \tab  \cr
   End \tab tterm \tab  \tab  \cr
   Title \tab treatyname \tab Common.Name \tab DocumentName \cr
   Memberships \tab  \tab StateName \tab Signatories \cr
   tfddID \tab  \tab  \tab 2016Update ID \cr
   mitchID \tab ieadbID \tab  \tab  \cr
   AgreementType \tab  \tab  \tab  \cr
   DocType \tab inclusion \tab  \tab  \cr
}
}

\keyword{datasets}
