% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytreaties-parties.R
\docType{data}
\name{parties}
\alias{parties}
\title{parties datacube}
\format{
\describe{
\item{HUGGO \if{html}{\out{<img src="https://mirrors.creativecommons.org/presskit/buttons/88x31/png/by.png" height="12"/>}}:}{A dataset with 91,673
observations and 30 variables:
treatyID, stateID, Title, Begin, End, StateSignature, StateRatification, StateForce, StateEnd, Rat=Notif, Accession, Succession, Signature, Force, url, StateForce2, StateForce3, StateEnd2, ..., match, and Coder.}
\item{IEADB: }{A dataset with 15,463
observations and 11 variables:
stateID, treatyID, Begin, End, manyID, Title, stateSignature, Signature, Rat, Force, and ieadbID.}
\item{DESTA: }{A dataset with 7,466
observations and 9 variables:
manyID, treatyID, Title, Begin, stateID, Signature, Force, StateName, and destaID.
\item{GPTAD: }{A dataset with 2,192
observations and 7 variables:
stateID, treatyID, Begin, manyID, Title, Signature, and Force.}
\item{TFDD: }{A dataset with 1,832
observations and 7 variables:
stateID, treatyID, Begin, manyID, Title, Signature, and tfddID.}
}
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
}
\source{
\itemize{
\item Hollway, James, Henrique Sposito, and Jael Tan. 2021. International agreements for manydata.
\item Mitchell, Ron B. et al. 2020. 'What we know (and could know) about international environmental agreements'. \emph{Global Environmental Politics} 20.1, pp. 103-121.
\item Dür, Andreas, Leonardo Baccini, and Manfred Elsig. 2014. 'The Design of International Trade Agreements: Introducing a New Database'. \emph{The Review of International Organizations} 9.3, pp. 353-375.
\item World Bank Group. 2014. \emph{Global Preferential Trade Agreement Database (GPTAD)}. Online database. World Bank Group.
\item Oregon College of Earth and Oregon State University Atmospheric Science. 2021. \emph{Product of the Transboundary Freshwater Dispute Database}.
}
}
\usage{
parties
}
\description{
The \code{manytreaties::parties} datacube is a list containing 5 datasets: HUGGO, IEADB, GPTAD, TFDD, and DESTA.
It is a work-in-progress, so please let us know if you have any comments or suggestions.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $HUGGO
#> ------------------------------------------------------------
#> |    Variable     |  Class  |  Obs  |  Missing  |  Miss \%  |
#> ------------------------------------------------------------
#> |treatyID         |character|  91673|          0|         0|
#> |stateID          |character|  91673|          0|         0|
#> |Title            |character|  91673|          0|         0|
#> |Begin            |mdate    |  91661|         12|      0.01|
#> |End              |mdate    |  52885|      38788|     42.31|
#> |StateSignature   |mdate    |  73756|      17917|     19.54|
#> |StateRatification|mdate    |  27769|      63904|     69.71|
#> |StateForce       |mdate    |  40826|      50847|     55.47|
#> |StateEnd         |mdate    |  58205|      33468|     36.51|
#> |Rat=Notif        |character|   1227|      90446|     98.66|
#> |Accession        |mdate    |  14472|      77201|     84.21|
#> |Succession       |character|  91673|          0|         0|
#> |Signature        |mdate    |  91657|         16|      0.02|
#> |Force            |mdate    |  57124|      34549|     37.69|
#> |url              |character|    931|      90742|     98.98|
#> |StateForce2      |character|   6687|      84986|     92.71|
#> |StateForce3      |character|   3690|      87983|     95.97|
#> |StateEnd2        |character|   3696|      87977|     95.97|
#> |Term             |character|   1178|      90495|     98.71|
#> |Comments         |character|   3280|      88393|     96.42|
#> |obsolete         |character|  12575|      79098|     86.28|
#> |ProvisionalApp   |character|    772|      90901|     99.16|
#> |Reservation      |character|    756|      90917|     99.18|
#> |Notes            |character|    186|      91487|      99.8|
#> |stateForce_ecolex|character|  18517|      73156|      79.8|
#> |stateForce_iea   |character|  18517|      73156|      79.8|
#> |Consent          |character|    224|      91449|     99.76|
#> |Acceptance       |character|    525|      91148|     99.43|
#> |match            |character|      4|      91669|       100|
#> |Coder            |character|  91658|         15|      0.02|
#> ------------------------------------------------------------
#> 
#> 
#> $IEADB
#> ---------------------------------------------------------
#> |   Variable   |  Class  |  Obs  |  Missing  |  Miss \%  |
#> ---------------------------------------------------------
#> |stateID       |character|  15463|          0|         0|
#> |treatyID      |character|  15463|          0|         0|
#> |Begin         |mdate    |  15463|          0|         0|
#> |End           |mdate    |    870|      14593|     94.37|
#> |manyID        |character|  15463|          0|         0|
#> |Title         |character|  15463|          0|         0|
#> |stateSignature|mdate    |   9666|       5797|     37.49|
#> |Signature     |mdate    |  15463|          0|         0|
#> |Rat           |mdate    |   9978|       5485|     35.47|
#> |Force         |mdate    |  15463|          0|         0|
#> |ieadbID       |numeric  |  15463|          0|         0|
#> ---------------------------------------------------------
#> 
#> 
#> $GPTAD
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |stateID     |character|   2137|         55|      2.51|
#> |treatyID    |character|   2192|          0|         0|
#> |Begin       |mdate    |   2093|         99|      4.52|
#> |manyID      |character|   2192|          0|         0|
#> |Title       |character|   2192|          0|         0|
#> |Signature   |mdate    |   1882|        310|     14.14|
#> |Force       |mdate    |   1907|        285|        13|
#> -------------------------------------------------------
#> 
#> 
#> $TFDD
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |stateID     |character|   1831|          1|      0.05|
#> |treatyID    |character|   1832|          0|         0|
#> |Begin       |mdate    |   1730|        102|      5.57|
#> |manyID      |character|   1832|          0|         0|
#> |Title       |character|   1831|          1|      0.05|
#> |Signature   |mdate    |   1730|        102|      5.57|
#> |tfddID      |numeric  |   1831|          1|      0.05|
#> -------------------------------------------------------
#> 
#> 
#> $DESTA
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |manyID      |character|   7466|          0|         0|
#> |treatyID    |character|   7466|          0|         0|
#> |Title       |character|   7466|          0|         0|
#> |Begin       |mdate    |   6836|        630|      8.44|
#> |stateID     |character|   7455|         11|      0.15|
#> |Signature   |mdate    |   6836|        630|      8.44|
#> |Force       |mdate    |   4602|       2864|     38.36|
#> |StateName   |character|   7466|          0|         0|
#> |destaID     |character|   7466|          0|         0|
#> -------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{Mapping}{
\tabular{lllll}{
   \emph{manytreaties} \tab \emph{IEADB} \tab \emph{DESTA} \tab \emph{GPTAD} \tab \emph{TFDD} \cr
   stateID \tab country \tab Member \tab  \tab CCODE \cr
   Signature \tab csig \tab year \tab Date.of.Signature \tab DateSigned \cr
   Rat \tab crat \tab  \tab  \tab  \cr
   Force \tab ceif3 \tab entryforceyear \tab Date.of.Entry.into.Force \tab  \cr
   Force2 \tab ceif4 \tab  \tab  \tab  \cr
   End \tab tterm \tab  \tab  \tab  \cr
   Title \tab treatyname \tab name \tab Common.Name \tab DocumentName \cr
   Memberships \tab  \tab  \tab StateName \tab Signatories \cr
   destaID \tab  \tab base_treaty \tab  \tab  \cr
   tfddID \tab  \tab  \tab  \tab 2016Update ID \cr
   mitchID \tab ieadbID \tab  \tab  \tab  \cr
   DocType \tab inclusion \tab  \tab  \tab  \cr
}
}

\keyword{datasets}
